% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-topTables_extract.R
\name{addHoverText}
\alias{addHoverText}
\title{Add column with hover text to top table}
\usage{
addHoverText(input, columns)
}
\arguments{
\item{input}{top table}

\item{columns}{columns to extract from top table for hover info}
}
\value{
top table with additional 'hoverText' column
}
\description{
Add column with hover text to top table
}
\author{
Katarzyna Gorczak
}
