% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-topTables_arrange.R
\name{arrangeTopTables}
\alias{arrangeTopTables}
\title{Arrange top tables}
\usage{
arrangeTopTables(
  input,
  featuresIdVar,
  logFCrange = NULL,
  commonFeatures = FALSE,
  fdr = NULL,
  dir = NULL,
  output = c("table", "list")
)
}
\arguments{
\item{input}{list of top tables or long table}

\item{featuresIdVar}{column with feature identifier (must be unique)}

\item{logFCrange}{numeric vector with two values 
(low- and high-threshold for log-foldchange)}

\item{commonFeatures}{logical whether to keep the same set of features
per coefficients}

\item{fdr}{fdr threshold}

\item{dir}{direction for the logFC ("pos" or "neg"; NULL by default). 
The features with "pos" diraction will be extracted with logFC above 0. 
The features with "neg" diraction will be extracted with logFC below 0.}

\item{output}{whether to output list of top tables or long table}
}
\value{
arranged list or long data.frame with top tables for coefficients 
of interest
}
\description{
Arrange top tables
}
\author{
Katarzyna Gorczak
}
