% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-plot_functions.R
\name{calcCorrelation}
\alias{calcCorrelation}
\title{Calculate correlation}
\usage{
calcCorrelation(input)
}
\arguments{
\item{input}{long-format table}
}
\value{
text with information from \code{cor.test}
}
\description{
Calculate correlation
}
\author{
Katarzyna Gorczak
}
