% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots-genesBarplot.R
\name{callBarplot}
\alias{callBarplot}
\title{Create ggplot object with barplot}
\usage{
callBarplot(
  tbl,
  color,
  annotCex,
  addPercentage,
  title,
  titleCex,
  xlab,
  ylab,
  axesTitleCex,
  axesCex,
  legendPosition,
  legendTitleCex,
  legendCex
)
}
\arguments{
\item{tbl}{combined summaries for all coefficients 
(i.e.: number of significantly up- and down-regulated genes)}

\item{color}{color palette, must contain two colors.}

\item{annotCex}{cex for the text displayed in the plot 
(indicating number of genes).}

\item{addPercentage}{logical whether to add percentage}

\item{title}{plot title, NULL by default.}

\item{titleCex}{cex for the plot title.}

\item{xlab}{x-axis title, NULL by default.}

\item{ylab}{y-axis title, NULL by default.}

\item{axesTitleCex}{cex for the axis title.}

\item{axesCex}{cex for the axis text.}

\item{legendPosition}{legend position ("right", "bottom", "none"). If 'none',
no legend is shown.}

\item{legendTitleCex}{cex for the legend title.}

\item{legendCex}{cex for the legend text.}
}
\value{
ggplot object
}
\description{
Create ggplot object with barplot
}
\author{
Katarzyna Gorczak
}
