% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots-heatmap.R
\name{callHeatmap}
\alias{callHeatmap}
\title{Create ggplot object with heatmap}
\usage{
callHeatmap(
  topTableOutput,
  xlab,
  ylab,
  title,
  color,
  axesCex,
  featuresColor,
  axesTitleCex,
  titleCex,
  legendTitleCex,
  legendCex,
  typePlot,
  coefColor,
  colorNA
)
}
\arguments{
\item{topTableOutput}{combined topTables for all coefficients}

\item{xlab}{x-axis title, NULL by default.}

\item{ylab}{y-axis title, NULL by default.}

\item{title}{plot title, NULL by default.}

\item{color}{palette for gradient to fill the heatmap.}

\item{axesCex}{cex for the axis text.}

\item{featuresColor}{character vector specifying colors to use for 
the feature label text. Either length 1 or the same length as 
\code{features}, allowing particular features to be highlighted.}

\item{axesTitleCex}{cex for the axis title.}

\item{titleCex}{cex for the plot title.}

\item{legendTitleCex}{cex for the legend title.}

\item{legendCex}{cex for the legend text.}

\item{typePlot}{plot type, can be one of "static" or "interactive".}

\item{coefColor}{color palette for coef labels.}

\item{colorNA}{color for missing values.}
}
\value{
ggplot object
}
\description{
Create ggplot object with heatmap
}
\author{
Katarzyna Gorczak
}
