% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots-logRatioPlot.R
\name{callLogRatioPlot}
\alias{callLogRatioPlot}
\title{Create ggplot object with log ratio plot}
\usage{
callLogRatioPlot(
  topTableOutput,
  xlab,
  xexpand,
  axesCex,
  axesTitleCex,
  title,
  titleCex,
  facetNCol,
  facetCex,
  facetColor,
  color,
  textVar = NULL,
  textVarCex,
  featuresColor,
  typePlot,
  multiplePlot
)
}
\arguments{
\item{topTableOutput}{combined topTables for all coefficients}

\item{xlab}{x-axis title, or \code{NULL} to remove.\cr
By default, set to: 'logFC' or 'logFC (+- SE)' if error bar(s)
are available.}

\item{xexpand}{expansion factor for the x-axis 
(see \code{\link[ggplot2]{expansion}}. If \code{text} is specified, 
the x-axis is expanded by 20\% on each side.}

\item{axesCex}{relative size for the label of the axes.}

\item{axesTitleCex}{relative size for the title of the axes.}

\item{title}{plot title, NULL by default.}

\item{titleCex}{relative size for the title of the plot.}

\item{facetNCol}{number of columns in facets, by default the function 
\code{n2mfrow} is used to.}

\item{facetCex}{cex for the facets if multiple plot are used.}

\item{facetColor}{color for the text of the facets}

\item{color}{string with colors for the bars, one per coefficient.}

\item{textVar}{(optional) String with name of a column to display as text.}

\item{textVarCex}{cex for the text next to bars}

\item{featuresColor}{character vector specifying colors to use for 
the feature label text. Either length 1 or the same length as 
\code{features}, allowing particular features to be highlighted.}

\item{typePlot}{plot type, can be one of "static" or "interactive".}

\item{multiplePlot}{whether to use facet_wrap on coefficients}
}
\value{
ggplot object
}
\description{
Create ggplot object with log ratio plot
}
\author{
Laure Cougnaud, Kirsten Van Hoorde, Katarzyna Gorczak
}
