% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots-MAplot.R
\name{callMAplot}
\alias{callMAplot}
\title{Create ggplot object with MA plot}
\usage{
callMAplot(
  topTableOutput,
  topTableOutputTopGenes,
  includeTableGenesOfInterest,
  title,
  topTableOutputGenesOfInterest,
  xlab,
  ylab,
  axesCex,
  axesTitleCex,
  titleCex,
  facetCex,
  facetColor,
  facetNCol,
  genesToHighlightCex,
  topGenesCex,
  color,
  direction,
  alpha,
  sizeVar,
  size,
  legendPosition,
  legendTitleCex,
  legendCex,
  typePlot,
  multiplePlot,
  ...
)
}
\arguments{
\item{topTableOutput}{combined topTables for all coefficients}

\item{topTableOutputTopGenes}{data.frame with top genes}

\item{includeTableGenesOfInterest}{whether to label \code{genesToHighlight}}

\item{title}{plot title, NULL by default.}

\item{topTableOutputGenesOfInterest}{data.frame with \code{genesToHighlight}}

\item{xlab}{x-axis title, NULL by default.}

\item{ylab}{y-axis title, NULL by default.}

\item{axesCex}{cex for the axis text.}

\item{axesTitleCex}{cex for the axis title.}

\item{titleCex}{cex for the plot title.}

\item{facetCex}{cex for the facets if multiple plot are used.}

\item{facetColor}{color for the text of the facets}

\item{facetNCol}{number of columns in facets, by default the function 
\code{n2mfrow} is used to.}

\item{genesToHighlightCex}{cex for genesToHighlight}

\item{topGenesCex}{cex for topGenes labels}

\item{color}{colors for points indicating direction (should be three colors:
significant up- and down-regulated and non-significant)}

\item{direction}{logical whether to color significant up- and 
down-regulated genes}

\item{alpha}{transparency level for the points, 0.4 by default.}

\item{sizeVar}{column name used for the size, empty by default.}

\item{size}{character or factor with specified size(s) (cex) for the points, 
replicated if needed. This is used only if sizeVar is empty. 
By default: '2.5' if sizeVar is not specified and 
default ggplot size(s) otherwise}

\item{legendPosition}{legend position ("right", "bottom", "none"). 
If 'none', no legend is shown.}

\item{legendTitleCex}{cex for the legend title.}

\item{legendCex}{cex for the legend text.}

\item{typePlot}{plot type can be one of "static" or "interactive".}

\item{multiplePlot}{whether to use facet_wrap on coefficients}

\item{...}{Extra parameters passed to \code{geom_text_repel} to customize
the position of the gene labels.}
}
\value{
ggplot object
}
\description{
Create ggplot object with MA plot
}
\author{
Katarzyna Gorczak
}
