% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots-scatterPlot.R
\name{callScatterPlot}
\alias{callScatterPlot}
\title{Create ggplot object with scatter plot}
\usage{
callScatterPlot(
  topTableOutput,
  topTableOutputTopGenes,
  includeTableGenesOfInterest,
  coef,
  topTableOutputGenesOfInterest,
  facetNCol,
  xlab,
  ylab,
  title,
  fdr,
  color,
  axesCex,
  axesTitleCex,
  titleCex,
  legendPosition,
  legendTitleCex,
  legendCex,
  facetCex,
  facetColor,
  topGenesCex,
  genesToHighlightCex,
  alpha,
  pointSize,
  typePlot,
  multiplePlot,
  correlation,
  correlationCex,
  ...
)
}
\arguments{
\item{topTableOutput}{combined topTables for all coefficients}

\item{topTableOutputTopGenes}{data.frame with top genes}

\item{includeTableGenesOfInterest}{whether to label \code{genesToHighlight}}

\item{coef}{character, coefficient names.}

\item{topTableOutputGenesOfInterest}{data.frame with \code{genesToHighlight}}

\item{facetNCol}{number of columns in facets, by default the function 
\code{n2mfrow} is used to.}

\item{xlab}{x-axis title, NULL by default.}

\item{ylab}{y-axis title, NULL by default.}

\item{title}{plot title, NULL by default.}

\item{fdr}{threshold considered for significance, NULL by default.}

\item{color}{color palette to distinguish significance groups. 
Four colors must be specified.}

\item{axesCex}{cex for the axis text.}

\item{axesTitleCex}{cex for the axis title.}

\item{titleCex}{cex for the plot title.}

\item{legendPosition}{legend position ("right", "bottom", "none"). 
If 'none', no legend is shown.}

\item{legendTitleCex}{cex for the legend title.}

\item{legendCex}{cex for the legend text.}

\item{facetCex}{cex for the facets if multiple plot are used.}

\item{facetColor}{color for the text of the facets}

\item{topGenesCex}{cex for topGenes labels}

\item{genesToHighlightCex}{cex for genesToHighlight}

\item{alpha}{transparency level for the points, 0.4 by default.}

\item{pointSize}{point size, 2 by default.}

\item{typePlot}{plot type, can be one of "static" or "interactive"}

\item{multiplePlot}{logical whether to use facet on coefficients}

\item{correlation}{logical whether to calculate the correlation and add 
text to the plot.}

\item{correlationCex}{cex for the text with correlation value}

\item{...}{Extra parameters passed to \code{geom_text_repel} to customize
the position of the gene labels.}
}
\value{
ggplot object
}
\description{
Create ggplot object with scatter plot
}
\author{
Katarzyna Gorczak
}
