% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots-volcanoPlot.R
\name{callVolcanoPlot}
\alias{callVolcanoPlot}
\title{Create ggplot object with volcano plot}
\usage{
callVolcanoPlot(
  topTableOutput,
  topTableOutputTopGenes,
  includeTableGenesOfInterest,
  topTableOutputGenesOfInterest,
  xlab,
  ylab,
  title,
  fdr,
  colorVar,
  color,
  shapeVar,
  shape,
  alphaVar,
  alpha,
  alphaRange,
  sizeVar,
  size,
  sizeRange,
  axesCex,
  axesTitleCex,
  titleCex,
  legendPosition,
  legendTitleCex,
  legendCex,
  facetCex,
  facetColor,
  facetNCol,
  topGenesCex,
  genesToHighlightCex,
  typePlot,
  multiplePlot,
  ...
)
}
\arguments{
\item{topTableOutput}{combined topTables for all coefficients}

\item{topTableOutputTopGenes}{data.frame with top genes}

\item{includeTableGenesOfInterest}{whether to label \code{genesToHighlight}}

\item{topTableOutputGenesOfInterest}{data.frame with \code{genesToHighlight}}

\item{xlab}{x-axis title, NULL by default.}

\item{ylab}{y-axis title, NULL by default.}

\item{title}{plot title, NULL by default.}

\item{fdr}{threshold considered for significance, NULL by default.}

\item{colorVar}{(optional) string with column name containing variable used 
for coloring. If not specified, coloring is based on the adjusted p-values.}

\item{color}{color palette, only used if \code{colorVar} is specified.}

\item{shapeVar}{(optional) string with column name containing variable used 
for shaping. If not specified, only points are used.}

\item{shape}{shape palette, only used if \code{shapeVar} is specified.}

\item{alphaVar}{column name used for the transparency, empty by default.}

\item{alpha}{character or factor with specified transparency(s) for the 
points, replicated if needed. By default: '1' if alphaVar is not specified.}

\item{alphaRange}{transparency (alpha) range used in the plot, possible only
if the alphaVar is 'numeric' or 'integer'.}

\item{sizeVar}{column name used for the size, empty by default.}

\item{size}{character or factor with specified size(s) (cex) for the points, 
replicated if needed. This is used only if sizeVar is empty. 
By default: '2.5' if sizeVar is not specified and 
default ggplot size(s) otherwise}

\item{sizeRange}{size (cex) range used in the plot, 
possible only if the sizeVar is 'numeric' or 'integer'}

\item{axesCex}{cex for the axis text.}

\item{axesTitleCex}{cex for the axis title.}

\item{titleCex}{cex for the plot title.}

\item{legendPosition}{legend position ("right", "bottom", "none"). 
If 'none', no legend is shown.}

\item{legendTitleCex}{cex for the legend title.}

\item{legendCex}{cex for the legend text.}

\item{facetCex}{cex for the facets if multiple plot are used.}

\item{facetColor}{color for the text of the facets}

\item{facetNCol}{number of columns in facets, by default the function 
\code{n2mfrow} is used to.}

\item{topGenesCex}{cex for topGenes labels}

\item{genesToHighlightCex}{cex for genesToHighlight}

\item{typePlot}{plot type can be one of "static" or "interactive".}

\item{multiplePlot}{whether to use facet_wrap on coefficients}

\item{...}{Extra parameters passed to \code{geom_text_repel} to customize
the position of the gene labels.}
}
\value{
ggplot object
}
\description{
Create ggplot object with volcano plot
}
\author{
Laure Cougnaud, Kirsten Van Hoorde, Katarzyna Gorczak
}
