% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots-waterfallPlot.R
\name{callWaterfallPlot}
\alias{callWaterfallPlot}
\title{Create ggplot object with waterfall plot}
\usage{
callWaterfallPlot(
  topTableOutput,
  featuresColor,
  xlab,
  ylab,
  axesCex,
  axesTitleCex,
  title,
  titleCex,
  facetNCol,
  fillVar,
  fill,
  colorVar,
  color,
  alphaVar,
  alpha,
  alphaRange,
  legendPosition,
  legendCex,
  legendTitleCex,
  facetCex,
  facetColor,
  typePlot,
  multiplePlot
)
}
\arguments{
\item{topTableOutput}{combined top tables for all coefficients}

\item{featuresColor}{character vector specifying colors to use for 
the feature label text. Either length 1 or the same length as 
\code{features}, allowing particular features to be highlighted.}

\item{xlab}{x-axis title, 'logFC' by default.}

\item{ylab}{y-axis title, NULL by default.}

\item{axesCex}{cex for the axis text.}

\item{axesTitleCex}{cex for the axis title.}

\item{title}{plot title, NULL by default.}

\item{titleCex}{cex for the plot title.}

\item{facetNCol}{number of columns in facets, by default the function 
\code{n2mfrow} is used to.}

\item{fillVar}{name of variable (in 'input') used for filling, 
empty by default}

\item{fill}{character or factor with specified color(s) for the boxplot 
inside, replicated if needed. By default: 'skyblue2' if fillVar is not 
specified and default ggplot palette otherwise.}

\item{colorVar}{name of variable (in 'input') used for coloring, 
empty by default}

\item{color}{character or factor with specified color(s) for the bar border, 
replicated if needed. By default: 'skyblue2' if colorVar is not specified 
and default ggplot palette otherwise.}

\item{alphaVar}{column name used for the transparency, empty by default.}

\item{alpha}{character or factor with specified transparency(s) for the bars,
replicated if needed. By default: '1' if alphaVar is not specified.}

\item{alphaRange}{transparency (alpha) range used in the plot, possible only
if the alphaVar is 'numeric' or 'integer'.}

\item{legendPosition}{legend position ("right", "bottom", "none"). 
If 'none', no legend is shown.}

\item{legendCex}{cex for the legend text.}

\item{legendTitleCex}{cex for the legend title.}

\item{facetCex}{cex for the facets if multiple plot are used.}

\item{facetColor}{color for the text of the facets}

\item{typePlot}{plot type, can be one of "static" or "interactive".}

\item{multiplePlot}{logical whether to use facet_wrap on coefficients}
}
\value{
ggplot object
}
\description{
Create ggplot object with waterfall plot
}
\author{
Katarzyna Gorczak
}
