% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-plot_functions.R
\name{cbindFill}
\alias{cbindFill}
\title{Combine data.frames by columns allowing for different number of rows.}
\usage{
cbindFill(list, featuresIdVar, sort = FALSE)
}
\arguments{
\item{list}{list with data frames which should be combined by columns}

\item{featuresIdVar}{columns with feature ids by which data frames should
be combined}

\item{sort}{logical, whether the rows should be sorted}
}
\value{
data.frame
}
\description{
Combine data.frames by columns allowing for different number of rows.
}
\author{
Katarzyna Gorczak
}
