% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-checkCoefficients.R
\name{checkCoef}
\alias{checkCoef}
\title{Check coefficients}
\usage{
checkCoef(input, coef = NULL)
}
\arguments{
\item{input}{model or a list with top tables named with coefficients.\cr
For model: object of class \code{MArrayLM} 
(linear model, see \code{\link[limma]{eBayes}}), \code{DGELRT} 
(see \code{edgeR}) or \code{DESeqResults} (see \code{DESeq2}) are supported.
See more details on the top table format in section:
\link[=daVis-common-doc]{Top table format}.}

\item{coef}{character vector with coefficients}
}
\value{
no returned value, error if any of the coefficients not present
}
\description{
Check coefficients
}
