% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-checkInputColumns.R
\name{checkColumns}
\alias{checkColumns}
\title{Check input fields}
\usage{
checkColumns(input, featuresIdVar, coef, cols, error = TRUE)
}
\arguments{
\item{input}{model or a list with top tables named with coefficients.\cr
For model: object of class \code{MArrayLM} 
(linear model, see \code{\link[limma]{eBayes}}), \code{DGELRT} 
(see \code{edgeR}) or \code{DESeqResults} (see \code{DESeq2}) are supported.
See more details on the top table format in section:
\link[=daVis-common-doc]{Top table format}.}

\item{featuresIdVar}{column name with unique feature identifiers}

\item{coef}{coefficient name}

\item{cols}{columns with feature annotation}

\item{error}{Logical, if TRUE (by default) an error is returned
if required columns are not present}
}
\value{
(invisibly) columns not available in the input and
error if required columns not present (if 
\code{error} is TRUE)
}
\description{
Check input fields
}
\author{
Katarzyna Gorczak
}
