% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-checkInput.R
\name{checkInput}
\alias{checkInput}
\title{Check input}
\usage{
checkInput(input, error = TRUE)
}
\arguments{
\item{input}{model or a list with top tables named with coefficients.\cr
For model: object of class \code{MArrayLM} 
(linear model, see \code{\link[limma]{eBayes}}), \code{DGELRT} 
(see \code{edgeR}) or \code{DESeqResults} (see \code{DESeq2}) are supported.
See more details on the top table format in section:
\link[=daVis-common-doc]{Top table format}.}

\item{error}{logical whether to return error message}
}
\value{
(invisible) logical indicating if \code{input} passes the check,
and error if input is not of the desired class (if \code{error} is TRUE).
}
\description{
Check input
}
