% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-plot_functions.R
\name{colorBlindPalette}
\alias{colorBlindPalette}
\title{Color-blind palette}
\usage{
colorBlindPalette(n, grey = TRUE)
}
\arguments{
\item{n}{number of colors to return}

\item{grey}{whether '#999999' color should be included}
}
\value{
n colors
}
\description{
Color-blind palette
}
