% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-plot_functions.R
\name{concatenateVars}
\alias{concatenateVars}
\title{Concatenate feature variables}
\usage{
concatenateVars(tbl, vars, nChar = NULL)
}
\arguments{
\item{tbl}{long-form top table for all coefficients}

\item{vars}{column names to concatenate}

\item{nChar}{maximum number of characters to truncate the 
feature labels to}
}
\value{
character vector with concatenated with ' | ' columns
}
\description{
Concatenate feature variables
}
\author{
Katarzyna Gorczak
}
