% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots-genesBarplot.R
\name{createDataBarplot}
\alias{createDataBarplot}
\title{Create table for barplot}
\usage{
createDataBarplot(
  input,
  coef,
  coefLabel = coef,
  fdr,
  logFCrange = NULL,
  addPercentage = FALSE
)
}
\arguments{
\item{input}{model or a list with top tables named with coefficients.\cr
For model: object of class \code{MArrayLM} 
(linear model, see \code{\link[limma]{eBayes}}), \code{DGELRT} 
(see \code{edgeR}) or \code{DESeqResults} (see \code{DESeq2}) are supported.
See more details on the top table format in section:
\link[=daVis-common-doc]{Top table format}.}

\item{coef}{character vector with coefficients}

\item{coefLabel}{character vector with coefficient labels or list with
such vector (in case of multiple labels)}

\item{fdr}{threshold considered for significance, NULL by default.}

\item{logFCrange}{numeric, two values (upper and lower bounds for logFC).}

\item{addPercentage}{logical whether to add percentage}
}
\value{
data.frame
}
\description{
Create table for barplot
}
\author{
Katarzyna Gorczak
}
