% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots-heatmap.R
\name{createDataHeatmap}
\alias{createDataHeatmap}
\title{Create table for heatmap}
\usage{
createDataHeatmap(
  input,
  coef,
  coefLabel = NULL,
  features = NULL,
  featuresIdVar,
  featuresVar,
  featuresMaxNChar,
  typePlot
)
}
\arguments{
\item{input}{model or a list with top tables named with coefficients.\cr
For model: object of class \code{MArrayLM} 
(linear model, see \code{\link[limma]{eBayes}}), \code{DGELRT} 
(see \code{edgeR}) or \code{DESeqResults} (see \code{DESeq2}) are supported.
See more details on the top table format in section:
\link[=daVis-common-doc]{Top table format}.}

\item{coef}{character, coefficient names.}

\item{coefLabel}{character vector of labels or a function to transform the 
existing labels.}

\item{features}{IDs of features to show. If NULL (by default), 
show top 20 features.}

\item{featuresIdVar}{column name with feature ids.}

\item{featuresVar}{column name with feature ids to label, the same as 
\code{featuresIdVar} by default.}

\item{featuresMaxNChar}{maximum number of characters to truncate the 
feature labels to.}

\item{typePlot}{plot type, can be one of "static" or "interactive".}
}
\value{
data.frame
}
\description{
Create table for heatmap
}
\author{
Laure Cougnaud, Katarzyna Gorczak
}
