% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots-logRatioPlot.R
\name{createDataLogRatioPlot}
\alias{createDataLogRatioPlot}
\title{Create table for log ratio plot}
\usage{
createDataLogRatioPlot(
  input,
  coef,
  coefLabel = NULL,
  text = NULL,
  features = NULL,
  featuresIdVar,
  featuresVar,
  featuresMaxNChar,
  order,
  featuresOrder,
  errorBars,
  typePlot
)
}
\arguments{
\item{input}{model or a list with top tables named with coefficients.\cr
For model: object of class \code{MArrayLM} 
(linear model, see \code{\link[limma]{eBayes}}), \code{DGELRT} 
(see \code{edgeR}) or \code{DESeqResults} (see \code{DESeq2}) are supported.
See more details on the top table format in section:
\link[=daVis-common-doc]{Top table format}.}

\item{coef}{character vector with coefficients}

\item{coefLabel}{character vector with coefficient labels or list with
such vector (in case of multiple labels)}

\item{text}{(optional) String with name of a column, or function to extract
a text from the columns of a top table, to display as text.\cr
See the available columns in the top table in section:
\link[=daVis-common-doc]{Top table format}.}

\item{features}{character vector with features to subset \code{input}. 
If NULL, top table for the top \code{n} features (for the first coefficient) 
is extracted.}

\item{featuresIdVar}{column name with unique feature ids (empty by default)}

\item{featuresVar}{column name with feature ids to label.}

\item{featuresMaxNChar}{maximum number of characters to truncate the 
feature labels to.}

\item{order}{logical whether to order features}

\item{featuresOrder}{if not NULL, features in the graph are re-ordered, 
either based on similarity ('similarity') or significance ('significance') 
of the statistics. See section 'Feature ordering'.}

\item{errorBars}{logical whether to add standard errors. 
Allowed only for \code{input} of class \code{"MArrayLM"} (or list of those).}

\item{typePlot}{plot type, can be one of "static" or "interactive".}
}
\value{
data.frame
}
\description{
Create table for log ratio plot
}
