% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots-MAplot.R
\name{createDataMAplot}
\alias{createDataMAplot}
\title{Create table for MA plot}
\usage{
createDataMAplot(
  input,
  coef,
  coefLabel = NULL,
  featuresIdVar,
  featuresVar,
  topGenes,
  topGenesVar,
  genesToHighlight,
  genesToHighlightVar,
  logFCrange,
  direction,
  sizeVar,
  fdr,
  typePlot
)
}
\arguments{
\item{input}{model or a list with top tables named with coefficients.\cr
For model: object of class \code{MArrayLM} 
(linear model, see \code{\link[limma]{eBayes}}), \code{DGELRT} 
(see \code{edgeR}) or \code{DESeqResults} (see \code{DESeq2}) are supported.
See more details on the top table format in section:
\link[=daVis-common-doc]{Top table format}.}

\item{coef}{character vector with coefficients}

\item{coefLabel}{character vector with coefficient labels or list with
such vector (in case of multiple labels)}

\item{featuresIdVar}{column name with unique feature ids (empty by default)}

\item{featuresVar}{column name with feature ids to label.}

\item{topGenes}{numeric, number of top genes with highest logFC or p-value 
to highlight in the plot for each considered coefficient, 0 by default.}

\item{topGenesVar}{column name with feature identifier to label topGenes}

\item{genesToHighlight}{string with identifiers of the genes to highlight, 
NULL by default. The gene identifiers should correspond to the variable 
specified in \code{featuresIdVar} or row names of \code{input}}

\item{genesToHighlightVar}{column name for the labels of genesToHighlight.
Same as \code{featuresIdVar} by default.}

\item{logFCrange}{numeric, two values (upper and lower bounds for logFC).}

\item{direction}{logical whether to color significant up- and 
down-regulated genes}

\item{sizeVar}{column name used for the size, empty by default.}

\item{fdr}{threshold considered for direction (up- or down- 
significant features), 0.05 by default.}

\item{typePlot}{plot type can be one of "static" or "interactive".}
}
\value{
data.frame
}
\description{
Create table for MA plot
}
\author{
Katarzyna Gorczak
}
