% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots-scatterPlot.R
\name{createDataScatterPlot}
\alias{createDataScatterPlot}
\title{Create table for the scatter plot}
\usage{
createDataScatterPlot(
  input,
  coef,
  coefLabel = NULL,
  featuresIdVar,
  featuresVar,
  fdr,
  topGenes,
  topGenesVar,
  genesToHighlight,
  genesToHighlightVar,
  genesToHighlightThresholdPValue,
  genesToHighlightThresholdLogFC,
  logFCrange,
  typePlot
)
}
\arguments{
\item{input}{model or a list with top tables named with coefficients.\cr
For model: object of class \code{MArrayLM} 
(linear model, see \code{\link[limma]{eBayes}}), \code{DGELRT} 
(see \code{edgeR}) or \code{DESeqResults} (see \code{DESeq2}) are supported.
See more details on the top table format in section:
\link[=daVis-common-doc]{Top table format}.}

\item{coef}{character vector with coefficients}

\item{coefLabel}{character vector with coefficient labels or list with
such vector (in case of multiple labels)}

\item{featuresIdVar}{column name with unique feature ids (empty by default)}

\item{featuresVar}{column name with feature ids to label, the same as 
'featuresIdVar' by default.}

\item{fdr}{threshold considered for significance, NULL by default.}

\item{topGenes}{numeric, number of top genes with highest logFC or p-value 
to highlight in the plot for each considered coefficient, 0 by default.}

\item{topGenesVar}{column name with feature identifier to label topGenes}

\item{genesToHighlight}{string with identifiers of the genes to highlight, 
NULL by default. The gene identifiers should correspond to the variable 
specified in \code{genesToHighlightVar}, and be contained among the column 
names of the output of the \code{topTable} function from \code{limma}.}

\item{genesToHighlightVar}{column name with the genes to highlight.
Same as \code{featuresIdVar} by default.}

\item{genesToHighlightThresholdPValue}{numeric, if specified (1 by default)
keep among the genes to highlight, the genes which have a raw p-value lower 
(strict) than this threshold for at least one of the coefficient considered.}

\item{genesToHighlightThresholdLogFC}{numeric, if specified (NULL by default)
keep among the genes to highlight, the genes which have an absolute log FC
higher (strict) than this threshold for at least one of the coefficient 
considered.}

\item{logFCrange}{numeric, two values (upper and lower bounds for logFC).}

\item{typePlot}{plot type, can be one of "static" or "interactive"}
}
\value{
data.frame
}
\description{
Create table for the scatter plot
}
\author{
Katarzyna Gorczak
}
