% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots-upsetPlot.R
\name{createDataUpsetPlot}
\alias{createDataUpsetPlot}
\title{Create data for upset plot}
\usage{
createDataUpsetPlot(
  input,
  coef,
  coefLabel = NULL,
  featuresIdVar = character(),
  fdr,
  dir
)
}
\arguments{
\item{input}{model or a list with top tables named with coefficients.\cr
For model: object of class \code{MArrayLM} 
(linear model, see \code{\link[limma]{eBayes}}), \code{DGELRT} 
(see \code{edgeR}) or \code{DESeqResults} (see \code{DESeq2}) are supported.
See more details on the top table format in section:
\link[=daVis-common-doc]{Top table format}.}

\item{coef}{character vector with coefficients}

\item{coefLabel}{character vector with coefficient labels or list with
such vector (in case of multiple labels)}

\item{featuresIdVar}{column name with unique feature ids (empty by default)}

\item{fdr}{threshold considered for significance, 0.05 by default.}

\item{dir}{direction for feature regulation ('up' to select up-regulated 
features; 'down' to select down-regulated features)}
}
\value{
data.frame compatible with UpSetR
}
\description{
Create data for upset plot
}
\author{
Kirsten Van Hoorde, Katarzyna Gorczak
}
