% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots-volcanoPlot.R
\name{createDataVolcanoPlot}
\alias{createDataVolcanoPlot}
\title{Create table for the volcano plot}
\usage{
createDataVolcanoPlot(
  input,
  coef,
  coefLabel = NULL,
  featuresIdVar,
  featuresVar,
  colorVar,
  shapeVar,
  alphaVar,
  sizeVar,
  fdr,
  additionalThresholdsAdjPValue,
  topGenes,
  topGenesVar,
  genesToHighlight,
  genesToHighlightVar,
  genesToHighlightThresholdPValue,
  genesToHighlightThresholdLogFC,
  logFCrange,
  typePlot
)
}
\arguments{
\item{input}{model or a list with top tables named with coefficients.\cr
For model: object of class \code{MArrayLM} 
(linear model, see \code{\link[limma]{eBayes}}), \code{DGELRT} 
(see \code{edgeR}) or \code{DESeqResults} (see \code{DESeq2}) are supported.
See more details on the top table format in section:
\link[=daVis-common-doc]{Top table format}.}

\item{coef}{character vector with coefficients}

\item{coefLabel}{character vector with coefficient labels or list with
such vector (in case of multiple labels)}

\item{featuresIdVar}{column name with unique feature ids (empty by default)}

\item{featuresVar}{column name with feature ids to label; used for extra 
labels if \code{typePlot} is 'interactive'.}

\item{colorVar}{(optional) string with column name containing variable used 
for coloring. If not specified, coloring is based on the adjusted p-values.}

\item{shapeVar}{(optional) string with column name containing variable used 
for shaping. If not specified, only points are used.}

\item{alphaVar}{column name used for the transparency, empty by default.}

\item{sizeVar}{column name used for the size, empty by default.}

\item{fdr}{threshold considered for significance, NULL by default.}

\item{additionalThresholdsAdjPValue}{numeric, additional adjusted p-values 
thresholds to use for the coloring of the points, and indicated 
in the legend.}

\item{topGenes}{numeric, number of top genes with highest logFC or p-value 
to highlight in the plot for each considered coefficient, 0 by default.}

\item{topGenesVar}{column name with feature identifier to label topGenes}

\item{genesToHighlight}{string with identifiers of the genes to highlight, 
NULL by default. The gene identifiers should correspond to the variable 
specified in \code{genesToHighlightVar}, and be contained among the column 
names of the output of the \code{topTable} function from \code{limma}.}

\item{genesToHighlightVar}{column name with the genes to highlight.
Same as \code{featuresIdVar} by default.}

\item{genesToHighlightThresholdPValue}{numeric, if specified (1 by default)
keep among the genes to highlight, the genes which have a raw p-value lower 
(strict) than this threshold for at least one of the coefficient considered.}

\item{genesToHighlightThresholdLogFC}{numeric, if specified (NULL by default)
keep among the genes to highlight, the genes which have an absolute log FC 
higher (strict) than this threshold for at least one of the coefficient 
considered.}

\item{logFCrange}{numeric, two values (upper and lower bounds for logFC).}

\item{typePlot}{plot type can be one of "static" or "interactive".}
}
\value{
data.frame
}
\description{
Create table for the volcano plot
}
\author{
Laure Cougnaud, Kirsten Van Hoorde, Katarzyna Gorczak, 
Michela Pasetto
}
