% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots-waterfallPlot.R
\name{createDataWaterfallPlot}
\alias{createDataWaterfallPlot}
\title{Create data for waterfall plot}
\usage{
createDataWaterfallPlot(
  input,
  coef,
  coefLabel = NULL,
  features = NULL,
  featuresIdVar,
  featuresVar,
  featuresMaxNChar,
  fillVar,
  colorVar,
  alphaVar,
  fdr,
  logFCrange,
  typePlot
)
}
\arguments{
\item{input}{model or a list with top tables named with coefficients.\cr
For model: object of class \code{MArrayLM} 
(linear model, see \code{\link[limma]{eBayes}}), \code{DGELRT} 
(see \code{edgeR}) or \code{DESeqResults} (see \code{DESeq2}) are supported.
See more details on the top table format in section:
\link[=daVis-common-doc]{Top table format}.}

\item{coef}{character vector with coefficients}

\item{coefLabel}{character vector with coefficient labels or list with
such vector (in case of multiple labels)}

\item{features}{character vector with features to subset \code{input}. 
If NULL, top table for the top \code{n} features (for the first coefficient) 
is extracted.}

\item{featuresIdVar}{column name with unique feature ids (empty by default)}

\item{featuresVar}{column name with feature ids to label, the same as 
'featuresIdVar' by default.}

\item{featuresMaxNChar}{numeric, maximum number of characters to truncate 
the feature labels to.}

\item{fillVar}{name of variable (in 'input') used for filling, 
empty by default}

\item{colorVar}{name of variable (in 'input') used for coloring, 
empty by default}

\item{alphaVar}{column name used for the transparency, empty by default.}

\item{fdr}{threshold considered for significance, NULL by default.}

\item{logFCrange}{numeric, two values (upper and lower bounds for logFC).}

\item{typePlot}{plot type, can be one of "static" or "interactive".}
}
\value{
data.frame
}
\description{
Create data for waterfall plot
}
\author{
Katarzyna Gorczak
}
