% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-exampleData.R
\name{createExampleData}
\alias{createExampleData}
\title{Create example limma-, edgeR-object or a list of top tables.}
\usage{
createExampleData(
  path = ".",
  output = c("limma", "topTable", "edgeR", "deseq2"),
  quiet = TRUE
)
}
\arguments{
\item{path}{absolute path where data should be downloaded to}

\item{output}{which object should be created. Possible options: 
limma, edgeR, topTableList, DESeq2}

\item{quiet}{if TRUE, suppress status messages (if any), and the progress bar
(see \code{download.file} function)}
}
\value{
list with specific objects (output from \code{limma}, \code{edgeR},
\code{topTable} and/or \code{DESeq2})
}
\description{
Create example limma-, edgeR-object or a list of top tables.
}
\examples{
tmpDir <- tempfile(); dir.create(tmpDir)
getData <- createExampleData(path = tmpDir, output = "limma")

}
