% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-exampleData.R
\name{createExpressionSet}
\alias{createExpressionSet}
\title{Create ExpressionSet object from example data}
\usage{
createExpressionSet(path)
}
\arguments{
\item{path}{absolute path where data should be downloaded to}
}
\value{
ExpressionSet object
}
\description{
Create ExpressionSet object from example data
}
\author{
Katarzyna Gorczak
}
