% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots-scatterPlot.R
\name{createPairData}
\alias{createPairData}
\title{Create pairwise data}
\usage{
createPairData(tbl, featuresIdVar, fdr, typePlot, columns)
}
\arguments{
\item{tbl}{list of top tables}

\item{featuresIdVar}{column name with feature ids, empty by default. 
If specified and input is a model, featuresIdVar should be a column 
name in 'genes' slot.}

\item{fdr}{threshold considered for significance, NULL by default.}

\item{typePlot}{plot type, can be one of "static" or "interactive"}

\item{columns}{columns of interest from top table}
}
\value{
data.frame with reference and compared contrasts
}
\description{
Create pairwise data
}
\author{
Katarzyna Gorczak
}
