% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-exampleData.R
\name{createSampleAnnotation}
\alias{createSampleAnnotation}
\title{Create sample annotation from example data}
\usage{
createSampleAnnotation(path)
}
\arguments{
\item{path}{absolute path where data should be downloaded to}
}
\value{
data.frame with sample annotation
}
\description{
Create sample annotation from example data
}
\author{
Katarzyna Gorczak
}
