% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-plot_functions.R
\name{createTopTableGenesOfInterest}
\alias{createTopTableGenesOfInterest}
\title{Create table for genes of interest}
\usage{
createTopTableGenesOfInterest(
  input,
  featuresIdVar = character(),
  genesToHighlight,
  genesToHighlightVar,
  genesToHighlightThresholdPValue = 1,
  genesToHighlightThresholdLogFC = NULL
)
}
\arguments{
\item{input}{long-format table with top tables for coefficients of interest}

\item{featuresIdVar}{column name with feature ids, 'ENTREZID' by default.}

\item{genesToHighlight}{string with identifiers of the genes to highlight.
The gene identifiers should correspond to the variable specified in 
\code{genesToHighlightVar}.}

\item{genesToHighlightVar}{column of the data to which the genes to 
highlight should be map}

\item{genesToHighlightThresholdPValue}{numeric, if specified (1 by default)
keep among the genes to highlight, the genes which have a raw p-value lower 
(strict) than this threshold for at least one of the coefficient considered}

\item{genesToHighlightThresholdLogFC}{numeric, if specified (NULL by default)
keep among the genes to highlight, the genes which have an absolute log FC 
higher (strict) than this threshold for at least one of the coefficient 
considered}
}
\value{
long-format table
}
\description{
Create table for genes of interest
}
\author{
Laure Cougnaud, Kirsten Van Hoorde
}
