% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots-heatmap.R
\name{daHeatmapLogFC}
\alias{daHeatmapLogFC}
\title{Heatmap of log fold-changes}
\usage{
daHeatmapLogFC(
  input,
  coef = NULL,
  coefLabel = NULL,
  features = NULL,
  featuresIdVar = character(),
  featuresVar = featuresIdVar,
  featuresColor = "black",
  featuresMaxNChar = 50,
  xlab = NULL,
  ylab = NULL,
  axesCex = 1,
  axesTitleCex = 1,
  title = NULL,
  titleCex = 1,
  legendTitleCex = 1,
  legendCex = 0.8,
  coefColor = "black",
  color = c("#0072B2", "white", "#D55E00"),
  colorNA = "grey",
  typePlot = c("static", "interactive")
)
}
\arguments{
\item{input}{model or a list with top tables named with coefficients.\cr
For model: object of class \code{MArrayLM} 
(linear model, see \code{\link[limma]{eBayes}}), \code{DGELRT} 
(see \code{edgeR}) or \code{DESeqResults} (see \code{DESeq2}) are supported.
See more details on the top table format in section:
\link[=daVis-common-doc]{Top table format}.}

\item{coef}{character, coefficient names.}

\item{coefLabel}{character vector of labels or a function to transform the 
existing labels.}

\item{features}{IDs of features to show. If NULL (by default), 
show top 20 features.}

\item{featuresIdVar}{column name with feature ids.}

\item{featuresVar}{column name with feature ids to label, the same as 
\code{featuresIdVar} by default.}

\item{featuresColor}{character vector specifying colors to use for 
the feature label text. Either length 1 or the same length as 
\code{features}, allowing particular features to be highlighted.}

\item{featuresMaxNChar}{maximum number of characters to truncate the 
feature labels to.}

\item{xlab}{x-axis title, NULL by default.}

\item{ylab}{y-axis title, NULL by default.}

\item{axesCex}{cex for the axis text.}

\item{axesTitleCex}{cex for the axis title.}

\item{title}{plot title, NULL by default.}

\item{titleCex}{cex for the plot title.}

\item{legendTitleCex}{cex for the legend title.}

\item{legendCex}{cex for the legend text.}

\item{coefColor}{color palette for coef labels.}

\item{color}{palette for gradient to fill the heatmap.}

\item{colorNA}{color for missing values.}

\item{typePlot}{plot type, can be one of "static" or "interactive".}
}
\value{
ggplot object
}
\description{
This is a function to create a heatmap that represents logFC values for 
several coefficients.
}
\examples{
exampleData <- createExampleData(path = ".", output = c("limma", "topTable"))
model <- exampleData$limma
topTableList <- exampleData$topTable
coefs <- c("B.LvsP", "L.LvsP", "B.PvsV", "L.PvsV")

# Simple heatmap
daHeatmapLogFC(input = model, coef = coefs)

# Specify feature annotation
daHeatmapLogFC(input = model, coef = coefs, featuresIdVar = "ENTREZID", 
featuresVar = c("SYMBOL", "GENENAME"), featuresMaxNChar = 35)

# Color coefficient labels
daHeatmapLogFC(input = model, coef = coefs, 
coefLabel = c("A", "B", "C", "D"), 
coefColor = c("blue", "red", "blue", "red"))

# Specify different set of features
coefs <- c("B.LvsP", "L.LvsP", "B.PvsV", "L.PvsV", "A")
daHeatmapLogFC(
input = list(model, A = topTableList[["B.LvsP"]][c(seq_len(6), 9, 10), ]), 
coef = coefs)

# see vignette for other examples

}
\author{
Kirsten Van Hoorde, Laure Cougnaud and Katarzyna Gorczak
}
