% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots-MAplot.R
\name{daMAplot}
\alias{daMAplot}
\title{MA plot}
\usage{
daMAplot(
  input,
  coef = NULL,
  coefLabel = NULL,
  featuresIdVar = character(),
  featuresVar = featuresIdVar,
  logFCrange = NULL,
  xlab = "log2 mean expression",
  ylab = "logFC",
  axesCex = 1,
  axesTitleCex = 1,
  title = NULL,
  titleCex = 1,
  legendPosition = c("right", "bottom", "none"),
  legendTitleCex = 1,
  legendCex = 0.8,
  facetCex = 1.1,
  facetColor = "black",
  facetNCol = grDevices::n2mfrow(length(coef))[2],
  topGenes = 0,
  topGenesVar = featuresIdVar,
  topGenesCex = 2.5,
  returnTopGenes = FALSE,
  genesToHighlight = NULL,
  genesToHighlightVar = featuresIdVar,
  genesToHighlightCex = 2.5,
  direction = FALSE,
  fdr = 0.05,
  color = if (direction) character() else "grey",
  alpha = 0.5,
  sizeVar = character(),
  size = if (length(sizeVar) > 0) numeric() else 2,
  typePlot = c("static", "interactive"),
  ...
)
}
\arguments{
\item{input}{model or a list with top tables named with coefficients.\cr
For model: object of class \code{MArrayLM} 
(linear model, see \code{\link[limma]{eBayes}}), \code{DGELRT} 
(see \code{edgeR}) or \code{DESeqResults} (see \code{DESeq2}) are supported.
See more details on the top table format in section:
\link[=daVis-common-doc]{Top table format}.}

\item{coef}{character, coefficient names.}

\item{coefLabel}{character vector of labels or a function to transform 
existing labels.}

\item{featuresIdVar}{column name with feature ids.}

\item{featuresVar}{column name with feature ids to label.}

\item{logFCrange}{numeric, two values (upper and lower bounds for logFC).}

\item{xlab}{x-axis title, NULL by default.}

\item{ylab}{y-axis title, NULL by default.}

\item{axesCex}{cex for the axis text.}

\item{axesTitleCex}{cex for the axis title.}

\item{title}{plot title, NULL by default.}

\item{titleCex}{cex for the plot title.}

\item{legendPosition}{legend position ("right", "bottom", "none"). 
If 'none', no legend is shown.}

\item{legendTitleCex}{cex for the legend title.}

\item{legendCex}{cex for the legend text.}

\item{facetCex}{cex for the facets if multiple plot are used.}

\item{facetColor}{color for the text of the facets}

\item{facetNCol}{number of columns in facets, by default the function 
\code{n2mfrow} is used to.}

\item{topGenes}{numeric, number of top genes with highest logFC or p-value 
to highlight in the plot for each considered coefficient, 0 by default.}

\item{topGenesVar}{column name with feature identifier to label topGenes}

\item{topGenesCex}{cex for topGenes labels}

\item{returnTopGenes}{logical, if TRUE (FALSE by default), return a list 
with the top genes highlighted in the plot.}

\item{genesToHighlight}{string with identifiers of the genes to highlight, 
NULL by default. The gene identifiers should correspond to the variable 
specified in \code{featuresIdVar} or row names of \code{input}}

\item{genesToHighlightVar}{column name for the labels of genesToHighlight.
Same as \code{featuresIdVar} by default.}

\item{genesToHighlightCex}{cex for genesToHighlight}

\item{direction}{logical whether to color significant up- and 
down-regulated genes}

\item{fdr}{threshold considered for direction (up- or down- 
significant features), 0.05 by default.}

\item{color}{colors for points indicating direction (should be three colors:
significant up- and down-regulated and non-significant)}

\item{alpha}{transparency level for the points, 0.4 by default.}

\item{sizeVar}{column name used for the size, empty by default.}

\item{size}{character or factor with specified size(s) (cex) for the points, 
replicated if needed. This is used only if sizeVar is empty. 
By default: '2.5' if sizeVar is not specified and 
default ggplot size(s) otherwise}

\item{typePlot}{plot type can be one of "static" or "interactive".}

\item{...}{Extra parameters passed to \code{geom_text_repel} to customize
the position of the gene labels.}
}
\value{
ggplot object or a list with ggplot object and top genes highlighted 
in the MA plot (top 10 genes with highest significance and/or highest logFC)
\code{featuresVar} with names \code{featuresIdVar} 
(if \code{returnTopGenes} is set to TRUE)
}
\description{
This is a function to create a MA plot. When several coefficients are used, 
multiple plots side by side are returned. 
MAplot visualizes mean expression versus logFC for each gene.
}
\examples{
exampleData <- createExampleData(path = ".", output = "limma")
model <- exampleData$limma
coefs <- c("B.LvsP", "L.LvsP")

# Simple MA plot
daMAplot(input = model, coef = coefs[1])

# Color by significant direction
daMAplot(input = model, coef = coefs, coefLabel = c("A", "B"),
direction = TRUE, color = c("steelblue", "firebrick", "grey"), facetNCol = 2)

# see vignette for other examples

}
\author{
Katarzyna Gorczak
}
