% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots-scatterPlot.R
\name{daScatterPlot}
\alias{daScatterPlot}
\title{Scatter plot}
\usage{
daScatterPlot(
  input,
  coef = NULL,
  coefLabel = coef,
  featuresIdVar = character(),
  featuresVar = featuresIdVar,
  fdr = 0.05,
  logFCrange = NULL,
  xlab = NULL,
  ylab = NULL,
  axesCex = 1,
  axesTitleCex = 1,
  title = NULL,
  titleCex = 1,
  legendPosition = c("right", "bottom", "none"),
  legendTitleCex = 1,
  legendCex = 0.8,
  facetCex = 1,
  facetColor = "black",
  facetNCol = grDevices::n2mfrow(length(coef))[2],
  color = c("gray90", "darkgoldenrod1", "dodgerblue", "darkgreen"),
  topGenes = 0,
  topGenesVar = featuresIdVar,
  topGenesCex = 2.5,
  returnTopGenes = FALSE,
  genesToHighlight = NULL,
  genesToHighlightVar = featuresIdVar,
  genesToHighlightCex = 2.5,
  genesToHighlightThresholdPValue = 1,
  genesToHighlightThresholdLogFC = NULL,
  alpha = 0.4,
  pointSize = 2,
  correlation = FALSE,
  correlationCex = 3,
  typePlot = c("static", "interactive"),
  ...
)
}
\arguments{
\item{input}{model or a list with top tables named with coefficients.\cr
For model: object of class \code{MArrayLM} 
(linear model, see \code{\link[limma]{eBayes}}), \code{DGELRT} 
(see \code{edgeR}) or \code{DESeqResults} (see \code{DESeq2}) are supported.
See more details on the top table format in section:
\link[=daVis-common-doc]{Top table format}.}

\item{coef}{character, coefficient names.}

\item{coefLabel}{character vector of labels or a function to transform 
existing labels.}

\item{featuresIdVar}{column name with feature ids, empty by default. 
If specified and input is a model, featuresIdVar should be a column 
name in 'genes' slot.}

\item{featuresVar}{column name with feature ids to label, the same as 
'featuresIdVar' by default.}

\item{fdr}{threshold considered for significance, NULL by default.}

\item{logFCrange}{numeric, two values (upper and lower bounds for logFC).}

\item{xlab}{x-axis title, NULL by default.}

\item{ylab}{y-axis title, NULL by default.}

\item{axesCex}{cex for the axis text.}

\item{axesTitleCex}{cex for the axis title.}

\item{title}{plot title, NULL by default.}

\item{titleCex}{cex for the plot title.}

\item{legendPosition}{legend position ("right", "bottom", "none"). 
If 'none', no legend is shown.}

\item{legendTitleCex}{cex for the legend title.}

\item{legendCex}{cex for the legend text.}

\item{facetCex}{cex for the facets if multiple plot are used.}

\item{facetColor}{color for the text of the facets}

\item{facetNCol}{number of columns in facets, by default the function 
\code{n2mfrow} is used to.}

\item{color}{color palette to distinguish significance groups. 
Four colors must be specified.}

\item{topGenes}{numeric, number of top genes with highest logFC or p-value 
to highlight in the plot for each considered coefficient, 0 by default.}

\item{topGenesVar}{column name with feature identifier to label topGenes}

\item{topGenesCex}{cex for topGenes labels}

\item{returnTopGenes}{logical, if TRUE (FALSE by default), return a list 
with the top genes highlighted in the plot.}

\item{genesToHighlight}{string with identifiers of the genes to highlight, 
NULL by default. The gene identifiers should correspond to the variable 
specified in \code{genesToHighlightVar}, and be contained among the column 
names of the output of the \code{topTable} function from \code{limma}.}

\item{genesToHighlightVar}{column name with the genes to highlight.
Same as \code{featuresIdVar} by default.}

\item{genesToHighlightCex}{cex for genesToHighlight}

\item{genesToHighlightThresholdPValue}{numeric, if specified (1 by default)
keep among the genes to highlight, the genes which have a raw p-value lower 
(strict) than this threshold for at least one of the coefficient considered.}

\item{genesToHighlightThresholdLogFC}{numeric, if specified (NULL by default)
keep among the genes to highlight, the genes which have an absolute log FC
higher (strict) than this threshold for at least one of the coefficient 
considered.}

\item{alpha}{transparency level for the points, 0.4 by default.}

\item{pointSize}{point size, 2 by default.}

\item{correlation}{logical whether to calculate the correlation and add 
text to the plot.}

\item{correlationCex}{cex for the text with correlation value}

\item{typePlot}{plot type, can be one of "static" or "interactive"}

\item{...}{Extra parameters passed to \code{geom_text_repel} to customize
the position of the gene labels.}
}
\value{
ggplot object or a list with ggplot object and top genes 
highlighted in the scatter plot 
(top 10 genes with highest significance and/or highest logFC)
\code{featuresVar} with names \code{featuresIdVar} 
(if \code{returnTopGenes} is set to TRUE)
}
\description{
This is a function to create a scatter plot. When several coefficients are 
used, multiple plots side by side are returned.
}
\examples{
exampleData <- createExampleData(path = ".", output = "limma")
model <- exampleData$limma
coefs <- c("B.LvsP", "L.LvsP", "B.PvsV", "L.PvsV")

# Simple scatter plot
daScatterPlot(input = model, coef = coefs[c(1,2)])

# More coefficients
daScatterPlot(input = model, coef = coefs, coefLabel = c("A", "B", "C", "D"),
facetNCol = 3)

# LogFC range
daScatterPlot(input = model, coef = coefs, coefLabel = c("A", "B", "C", "D"),
facetNCol = 3, logFCrange = c(-2, 2))

# see vignette for other examples

}
\author{
Katarzyna Gorczak
}
