% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots-genesBarplot.R
\name{daSignificantGenesBarplot}
\alias{daSignificantGenesBarplot}
\title{Barplot with the number of significant genes}
\usage{
daSignificantGenesBarplot(
  input,
  coef,
  coefLabel = coef,
  fdr = 0.05,
  logFCrange = NULL,
  xlab = NULL,
  ylab = NULL,
  axesCex = 1,
  axesTitleCex = 1.1,
  title = NULL,
  titleCex = 1.1,
  legendPosition = c("right", "bottom", "none"),
  legendTitleCex = 1,
  legendCex = 0.8,
  color = c("#32a6d3", "#e52323"),
  annotCex = 3.5,
  addPercentage = FALSE
)
}
\arguments{
\item{input}{model or a list with top tables named with coefficients.\cr
For model: object of class \code{MArrayLM} 
(linear model, see \code{\link[limma]{eBayes}}), \code{DGELRT} 
(see \code{edgeR}) or \code{DESeqResults} (see \code{DESeq2}) are supported.
See more details on the top table format in section:
\link[=daVis-common-doc]{Top table format}.}

\item{coef}{character, coefficient names.}

\item{coefLabel}{character vector of labels or a function to transform 
the existing labels.}

\item{fdr}{threshold considered for significance, NULL by default.}

\item{logFCrange}{numeric, two values (upper and lower bounds for logFC).}

\item{xlab}{x-axis title, NULL by default.}

\item{ylab}{y-axis title, NULL by default.}

\item{axesCex}{cex for the axis text.}

\item{axesTitleCex}{cex for the axis title.}

\item{title}{plot title, NULL by default.}

\item{titleCex}{cex for the plot title.}

\item{legendPosition}{legend position ("right", "bottom", "none"). If 'none',
no legend is shown.}

\item{legendTitleCex}{cex for the legend title.}

\item{legendCex}{cex for the legend text.}

\item{color}{color palette, must contain two colors.}

\item{annotCex}{cex for the text displayed in the plot 
(indicating number of genes).}

\item{addPercentage}{logical whether to add percentage}
}
\value{
ggplot object
}
\description{
This is a function to create a barplot indicating the number of significant 
genes in each coefficient. The number of up- and down-regulated genes is 
shown.
}
\examples{
exampleData <- createExampleData(path = ".", output = "limma")
model <- exampleData$limma
coefs <- c("B.LvsP", "L.LvsP", "B.PvsV", "L.PvsV")
 
# Simple barplot
daSignificantGenesBarplot(input = model, coef = coefs)

# Add percentage of genes
daSignificantGenesBarplot(input = model, coef = coefs, 
coefLabel = c("A", "B", "C", "D"), addPercentage = TRUE)

# see vignette for other examples

}
\author{
Katarzyna Gorczak
}
