% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots-upsetPlot.R
\name{daUpset}
\alias{daUpset}
\title{Upset plot for up- or down-regulated genes}
\usage{
daUpset(
  input,
  coef = NULL,
  coefLabel = coef,
  featuresIdVar = character(),
  fdr = 0.05,
  dir = c("up", "down"),
  ylab = "Intersection Size",
  xlab = "Set Size",
  axesCex = c(1.2, 1.7),
  axesTitleCex = c(1.4, 1.9),
  barsCex = 1.5,
  returnAnalysis = FALSE
)
}
\arguments{
\item{input}{model or a list with top tables named with coefficients.\cr
For model: object of class \code{MArrayLM} 
(linear model, see \code{\link[limma]{eBayes}}), \code{DGELRT} 
(see \code{edgeR}) or \code{DESeqResults} (see \code{DESeq2}) are supported.
See more details on the top table format in section:
\link[=daVis-common-doc]{Top table format}.}

\item{coef}{character, coefficient names.}

\item{coefLabel}{character vector of labels or a function to transform 
existing labels.}

\item{featuresIdVar}{column name with unique feature ids, empty by default.}

\item{fdr}{threshold considered for significance, 0.05 by default.}

\item{dir}{direction for feature regulation ('up' to select up-regulated 
features; 'down' to select down-regulated features)}

\item{ylab}{y-axis title}

\item{xlab}{x-axis title}

\item{axesCex}{cex for the axis text. If two numbers provided, the first 
one is used for the x-axis and the second one for y-axis.}

\item{axesTitleCex}{cex for the axis title. If two numbers provided, 
the first one is used for the x-axis and the second one for y-axis.}

\item{barsCex}{cex for the counts above the bars.}

\item{returnAnalysis}{logical, if TRUE (FALSE by default), return also the 
output of the analysis (list with all overlapping sets based on 
\code{featuresIdVar}), otherwise only the plot object}
}
\value{
uspet plot; if \code{returnAnalysis} is TRUE, return a list with 
overlapping sets and plot
}
\description{
This is a function to create a customized upset plot for up- or 
down-regulated genes.
}
\examples{
exampleData <- createExampleData(path = ".", output = "limma")
model <- exampleData$limma
coefs <- c("B.LvsP", "L.LvsP", "B.PvsV", "L.PvsV")

# Significantly up-regulated genes
daUpset(input = model, coef = coefs, fdr = 0.05, dir = "up")
 
# see vignette for other examples

}
\author{
Kirsten Van Hoorde, Katarzyna Gorczak, Michela Pasetto
}
