% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/daVis-common-args.R
\name{daVis-common-args}
\alias{daVis-common-args}
\title{Common parameters for the functions of the \code{daVis} package.}
\arguments{
\item{input}{model or a list with top tables named with coefficients.\cr
For model: object of class \code{MArrayLM} 
(linear model, see \code{\link[limma]{eBayes}}), \code{DGELRT} 
(see \code{edgeR}) or \code{DESeqResults} (see \code{DESeq2}) are supported.
See more details on the top table format in section:
\link[=daVis-common-doc]{Top table format}.}
}
\value{
No returned value
}
\description{
Common parameters for the functions of the \code{daVis} package.
}
