% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/daVis-common-args.R
\name{daVis-common-doc}
\alias{daVis-common-doc}
\title{Common documentation for the \code{daVis} package.}
\value{
No returned value
}
\description{
Common documentation for the \code{daVis} package.
}
\section{Top table format}{

The top table extracted from a specified model, or specified by the user
contains at least the columns:
\itemize{
\item{'logFC': log fold change}
\item{'AveExpr': average expression}
\item{'P.Value' and 'adj.PVal': raw and (multiplicity correction) adjusted 
p-values
\itemize{
\item{For \code{edgeR} model: this is extracted from the columns: 
'PValue' and 'FDR' respectively.}
}}
\item{feature identifier as specified by: \code{featuresIdVar}}
}
}

