% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots-waterfallPlot.R
\name{daWaterfallPlot}
\alias{daWaterfallPlot}
\title{Waterfall plot}
\usage{
daWaterfallPlot(
  input,
  features = NULL,
  coef,
  coefLabel = coef,
  featuresIdVar = character(),
  featuresVar = featuresIdVar,
  featuresColor = "black",
  featuresMaxNChar = 60,
  fdr = NULL,
  logFCrange = NULL,
  xlab = "logFC",
  ylab = NULL,
  axesCex = 1,
  axesTitleCex = 1.1,
  title = NULL,
  titleCex = 1.1,
  legendPosition = c("right", "bottom", "none"),
  legendTitleCex = 1,
  legendCex = 0.8,
  facetCex = 1,
  facetColor = "black",
  facetNCol = grDevices::n2mfrow(length(coef))[2],
  fillVar = character(),
  fill = if (length(fillVar) > 0) character() else "skyblue2",
  colorVar = character(),
  color = if (length(colorVar) > 0) character() else "white",
  alphaVar = character(),
  alpha = if (length(alphaVar) > 0) 1 else numeric(),
  alphaRange = numeric(),
  typePlot = c("static", "interactive")
)
}
\arguments{
\item{input}{model or a list with top tables named with coefficients.\cr
For model: object of class \code{MArrayLM} 
(linear model, see \code{\link[limma]{eBayes}}), \code{DGELRT} 
(see \code{edgeR}) or \code{DESeqResults} (see \code{DESeq2}) are supported.
See more details on the top table format in section:
\link[=daVis-common-doc]{Top table format}.}

\item{features}{IDs of features to show. If NULL (by default), 
show top 20 features.}

\item{coef}{character, coefficient names.}

\item{coefLabel}{character vector of labels or a function to transform 
existing labels.}

\item{featuresIdVar}{column name with feature ids, empty by default. 
If specified and input is a model, featuresIdVar should be a column name 
in 'genes' slot.}

\item{featuresVar}{column name with feature ids to label, the same as 
'featuresIdVar' by default.}

\item{featuresColor}{character vector specifying colors to use for 
the feature label text. Either length 1 or the same length as 
\code{features}, allowing particular features to be highlighted.}

\item{featuresMaxNChar}{numeric, maximum number of characters to truncate 
the feature labels to.}

\item{fdr}{threshold considered for significance, NULL by default.}

\item{logFCrange}{numeric, two values (upper and lower bounds for logFC).}

\item{xlab}{x-axis title, 'logFC' by default.}

\item{ylab}{y-axis title, NULL by default.}

\item{axesCex}{cex for the axis text.}

\item{axesTitleCex}{cex for the axis title.}

\item{title}{plot title, NULL by default.}

\item{titleCex}{cex for the plot title.}

\item{legendPosition}{legend position ("right", "bottom", "none"). 
If 'none', no legend is shown.}

\item{legendTitleCex}{cex for the legend title.}

\item{legendCex}{cex for the legend text.}

\item{facetCex}{cex for the facets if multiple plot are used.}

\item{facetColor}{color for the text of the facets}

\item{facetNCol}{number of columns in facets, by default the function 
\code{n2mfrow} is used to.}

\item{fillVar}{name of variable (in 'input') used for filling, 
empty by default}

\item{fill}{character or factor with specified color(s) for the boxplot 
inside, replicated if needed. By default: 'skyblue2' if fillVar is not 
specified and default ggplot palette otherwise.}

\item{colorVar}{name of variable (in 'input') used for coloring, 
empty by default}

\item{color}{character or factor with specified color(s) for the bar border, 
replicated if needed. By default: 'skyblue2' if colorVar is not specified 
and default ggplot palette otherwise.}

\item{alphaVar}{column name used for the transparency, empty by default.}

\item{alpha}{character or factor with specified transparency(s) for the bars,
replicated if needed. By default: '1' if alphaVar is not specified.}

\item{alphaRange}{transparency (alpha) range used in the plot, possible only
if the alphaVar is 'numeric' or 'integer'.}

\item{typePlot}{plot type, can be one of "static" or "interactive".}
}
\value{
ggplot object
}
\description{
This is a function to create a waterfall plot. When several coefficients are 
used, multiple plots side by side are returned.
}
\examples{
exampleData <- createExampleData(path = ".", output = "limma")
model <- exampleData$limma

# Simple waterfall plot
daWaterfallPlot(input = model, coef = "B.LvsP")

# see vignette for other examples

}
\author{
Katarzyna Gorczak
}
