% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-exampleData.R
\name{downloadData}
\alias{downloadData}
\title{Download example data from 
https://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSE60450}
\usage{
downloadData(path, quiet)
}
\arguments{
\item{path}{absolute path where data should be downloaded to}

\item{quiet}{if TRUE, suppress status messages (if any), and the progress bar
(see \code{download.file} function)}
}
\value{
(invisibly) downloads files at specific location ('path')
}
\description{
Download example data from 
https://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSE60450
}
\author{
Katarzyna Gorczak
}
