% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-topTables_extract.R
\name{extractColsOfInterest}
\alias{extractColsOfInterest}
\title{Extract columns of interest from top table}
\usage{
extractColsOfInterest(
  mean,
  topTableCoef,
  columns,
  stat,
  errorBars,
  coefI,
  text
)
}
\arguments{
\item{mean}{logical (FALSE by default) whether to extract column with 
averaged expression or logCPM ("AveExpr" or "logCPM") depending on the 
\code{input}}

\item{topTableCoef}{top table for one coefficient}

\item{columns}{columns to extract}

\item{stat}{logical (FALSE by default) whether to extract column with 
statistic: 't' or 'F'}

\item{errorBars}{logical whether to add standard errors. 
Allowed only for \code{input} of class \code{"MArrayLM"} (or list of those).}

\item{coefI}{coefficient of interest}

\item{text}{(optional) String with name of a column, or function to extract
a text from the columns of a top table, to display as text.\cr
See the available columns in the top table in section:
\link[=daVis-common-doc]{Top table format}.}
}
\value{
character vector
}
\description{
Extract columns of interest from top table
}
\author{
Laure Cougnaud, Katarzyna Gorczak
}
