% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots-upsetPlot.R
\name{extractFeatures}
\alias{extractFeatures}
\title{Extract feature ids for each overlapping set}
\usage{
extractFeatures(data)
}
\arguments{
\item{data}{data.frame with 0s and 1s; 
the number of columns corresponds to the number of coefs and the number of 
rows corresponds to the number of unique feature ids
each column is filled with 0s or 1s depending on the overlapping set}
}
\value{
list with features in overlapping sets
}
\description{
Extract feature ids for each overlapping set
}
\author{
Katarzyna Gorczak
}
