% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots-scatterPlot.R
\name{extractPairs}
\alias{extractPairs}
\title{Get top table per pair (ref and comp coef) for scatter plot}
\usage{
extractPairs(
  decideTestsOutput,
  refCoef,
  iCoef,
  multiplePlot,
  compCoef,
  tbl,
  typePlot,
  columns
)
}
\arguments{
\item{decideTestsOutput}{output of \code{decideTests}}

\item{refCoef}{coefficient on the x-axis}

\item{iCoef}{coefficients on the y-axis}

\item{multiplePlot}{whether to facet on coefs to compare}

\item{compCoef}{coefficients to compare with \code{refCoef}}

\item{tbl}{list of top tables}

\item{typePlot}{plot type, can be one of "static" or "interactive"}

\item{columns}{columns of interest from top table}
}
\value{
ggplot object
}
\description{
Get top table per pair (ref and comp coef) for scatter plot
}
\author{
Katarzyna Gorczak
}
