% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots-upsetPlot.R
\name{extractQueryList}
\alias{extractQueryList}
\title{Extract query list}
\usage{
extractQueryList(dataFrameSG)
}
\arguments{
\item{dataFrameSG}{data.frame compatible with UpSetR}
}
\value{
a list with queries for upset plot
}
\description{
Extract query list
}
