% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-topTables_extract.R
\name{extractTopTables}
\alias{extractTopTables}
\title{Extract top tables for coefficients of interest}
\usage{
extractTopTables(
  input,
  coef,
  coefLabel = NULL,
  features = NULL,
  n = 20,
  columns,
  errorBars = FALSE,
  mean = FALSE,
  stat = FALSE,
  featuresIdVar = character(),
  hoverText = FALSE,
  text = NULL,
  output = c("table", "list")
)
}
\arguments{
\item{input}{model or a list with top tables named with coefficients.\cr
For model: object of class \code{MArrayLM} 
(linear model, see \code{\link[limma]{eBayes}}), \code{DGELRT} 
(see \code{edgeR}) or \code{DESeqResults} (see \code{DESeq2}) are supported.
See more details on the top table format in section:
\link[=daVis-common-doc]{Top table format}.}

\item{coef}{character vector with coefficients}

\item{coefLabel}{character vector with coefficient labels or list with
such vector (in case of multiple labels)}

\item{features}{character vector with features to subset \code{input}. 
If NULL, top table for the top \code{n} features (for the first coefficient) 
is extracted.}

\item{n}{top features to extract. If set to Inf, all features are extracted}

\item{columns}{columns to extract}

\item{errorBars}{logical whether to add standard errors. 
Allowed only for \code{input} of class \code{"MArrayLM"} (or list of those).}

\item{mean}{logical (FALSE by default) whether to extract column with 
averaged expression or logCPM ("AveExpr" or "logCPM") depending on the 
\code{input}}

\item{stat}{logical (FALSE by default) whether to extract column with 
statistic: 't' or 'F'}

\item{featuresIdVar}{column name with unique feature ids (empty by default)}

\item{hoverText}{logical whether to add point label for interactive plot
(FALSE by default)}

\item{text}{(optional) String with name of a column, or function to extract
a text from the columns of a top table, to display as text.\cr
See the available columns in the top table in section:
\link[=daVis-common-doc]{Top table format}.}

\item{output}{output can be one of "table" or "list"}
}
\value{
data.frame with top tables or list of those 
(if \code{output} is 'list') for coefficients of interest
}
\description{
Extract top tables for coefficients of interest
}
\author{
Katarzyna Gorczak, Laure Cougnaud
}
