% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-topTables_extract.R
\name{extractTTcoef}
\alias{extractTTcoef}
\title{Extract top table for a specific coef}
\usage{
extractTTcoef(input, coefI, errorBars, text)
}
\arguments{
\item{input}{model or a list with top tables named with coefficients.\cr
For model: object of class \code{MArrayLM} 
(linear model, see \code{\link[limma]{eBayes}}), \code{DGELRT} 
(see \code{edgeR}) or \code{DESeqResults} (see \code{DESeq2}) are supported.
See more details on the top table format in section:
\link[=daVis-common-doc]{Top table format}.}

\item{coefI}{coefficient of interest}

\item{errorBars}{logical whether to add standard errors. 
Allowed only for \code{input} of class \code{"MArrayLM"} (or list of those).}

\item{text}{(optional) String with name of a column, or function to extract
a text from the columns of a top table, to display as text.\cr
See the available columns in the top table in section:
\link[=daVis-common-doc]{Top table format}.}
}
\value{
data.frame with statistics for a specific coef
}
\description{
Extract top table for a specific coef
}
\author{
Laure Cougnaud, Katarzyna Gorczak
}
