% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-plot_functions.R
\name{facet}
\alias{facet}
\title{Add facets to ggplot object}
\usage{
facet(g, topTableOutput, facetNCol, scales = "fixed")
}
\arguments{
\item{g}{ggplot object}

\item{topTableOutput}{combined topTables for all coefficients}

\item{facetNCol}{number of columns in facets, by default the function 
\code{n2mfrow} is used}

\item{scales}{see \code{facet_wrap} or \code{facet_nested_wrap}}
}
\value{
ggplot object
}
\description{
Add facets to ggplot object
}
\author{
Laure Cougnaud, Katarzyna Gorczak
}
