% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-topTables_arrange.R
\name{filterCommonFeatures}
\alias{filterCommonFeatures}
\title{Filter top tables based on common features across all tables}
\usage{
filterCommonFeatures(input, featuresIdVar)
}
\arguments{
\item{input}{list of top tables}

\item{featuresIdVar}{column with feature identifier (must be unique)}
}
\value{
list of top tables with common features
}
\description{
Filter top tables based on common features across all tables
}
\author{
Katarzyna Gorczak
}
