% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-plot_functions.R
\name{filterGenesOfInterest}
\alias{filterGenesOfInterest}
\title{Filter genes of interest}
\usage{
filterGenesOfInterest(
  input,
  featuresIdVar = character(),
  genesToHighlightVar,
  genesToHighlightThresholdPValue = 1,
  genesToHighlightThresholdLogFC = NULL
)
}
\arguments{
\item{input}{long-format table}

\item{featuresIdVar}{column name with feature ids, 'ENTREZID' by default.}

\item{genesToHighlightVar}{column name with genes to highlight}

\item{genesToHighlightThresholdPValue}{numeric, if specified 
keep among the genes to highlight, the genes which have a raw p-value lower 
(strict) than this threshold for at least one of the coefficient considered}

\item{genesToHighlightThresholdLogFC}{numeric, if specified 
keep among the genes to highlight, the genes which have an absolute log FC 
higher (strict) than this threshold for at least one of the coefficient 
considered}
}
\value{
long-format table
}
\description{
Filter genes of interest
}
\author{
Laure Cougnaud, Kirsten Van Hoorde, Katarzyna Gorczak
}
