% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-topTables_arrange.R
\name{filterLogFC}
\alias{filterLogFC}
\title{Filter top tables based on logFC threshold}
\usage{
filterLogFC(input, logFCrange)
}
\arguments{
\item{input}{list of top tables}

\item{logFCrange}{numeric vector with two values 
(low- and high-threshold for log-foldchange)}
}
\value{
list of top tables with subset of features based on logFC range
}
\description{
Filter top tables based on logFC threshold
}
\author{
Katarzyna Gorczak
}
