% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots-MAplot.R
\name{formatAesMA}
\alias{formatAesMA}
\title{Format aesthetics for the main plot with MA plot}
\usage{
formatAesMA(g, topTableOutput, direction, color, sizeVar, size)
}
\arguments{
\item{g}{ggplot object with MA plot}

\item{topTableOutput}{combined topTables for all coefficients}

\item{direction}{logical whether to color significant up- and 
down-regulated genes}

\item{color}{colors for points indicating direction (should be three colors:
significant up- and down-regulated and non-significant)}

\item{sizeVar}{column name used for the size, empty by default.}

\item{size}{character or factor with specified size(s) (cex) for the points, 
replicated if needed. This is used only if sizeVar is empty. 
By default: '2.5' if sizeVar is not specified and 
default ggplot size(s) otherwise}
}
\value{
ggplot object
}
\description{
Format aesthetics for the main plot with MA plot
}
\author{
Katarzyna Gorczak
}
