% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots-volcanoPlot.R
\name{formatAesVP}
\alias{formatAesVP}
\title{Format aesthetics for the main plot with volcano plot}
\usage{
formatAesVP(
  g,
  topTableOutput,
  colorVar,
  color,
  shapeVar,
  shape,
  alphaVar,
  alpha,
  alphaRange,
  sizeVar,
  size,
  sizeRange
)
}
\arguments{
\item{g}{ggplot object with volcano plot}

\item{topTableOutput}{combined topTables for all coefficients}

\item{colorVar}{(optional) string with column name containing variable used 
for coloring. If not specified, coloring is based on the adjusted p-values.}

\item{color}{color palette, only used if \code{colorVar} is specified.}

\item{shapeVar}{(optional) string with column name containing variable used 
for shaping. If not specified, only points are used.}

\item{shape}{shape palette, only used if \code{shapeVar} is specified.}

\item{alphaVar}{column name used for the transparency, empty by default.}

\item{alpha}{character or factor with specified transparency(s) for the 
points, replicated if needed. By default: '1' if alphaVar is not specified.}

\item{alphaRange}{transparency (alpha) range used in the plot, possible only
if the alphaVar is 'numeric' or 'integer'.}

\item{sizeVar}{column name used for the size, empty by default.}

\item{size}{character or factor with specified size(s) (cex) for the points, 
replicated if needed. This is used only if sizeVar is empty. 
By default: '2.5' if sizeVar is not specified and 
default ggplot size(s) otherwise}

\item{sizeRange}{size (cex) range used in the plot, 
possible only if the sizeVar is 'numeric' or 'integer'}
}
\value{
ggplot object
}
\description{
Format aesthetics for the main plot with volcano plot
}
\author{
Laure Cougnaud, Kirsten Van Hoorde, Katarzyna Gorczak
}
