% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots-waterfallPlot.R
\name{formatAesWP}
\alias{formatAesWP}
\title{Format aesthetics for the main plot with waterfall plot}
\usage{
formatAesWP(
  topTableOutput,
  colorVar,
  color,
  g,
  fillVar,
  fill,
  alphaVar,
  alpha,
  alphaRange
)
}
\arguments{
\item{topTableOutput}{combined topTables for all coefficients}

\item{colorVar}{name of variable (in 'input') used for coloring, 
empty by default}

\item{color}{character or factor with specified color(s) for the bar border, 
replicated if needed. By default: 'skyblue2' if colorVar is not specified 
and default ggplot palette otherwise.}

\item{g}{ggplot object with volcano plot}

\item{fillVar}{name of variable (in 'input') used for filling, 
empty by default}

\item{fill}{character or factor with specified color(s) for the boxplot 
inside, replicated if needed. By default: 'skyblue2' if fillVar is not 
specified and default ggplot palette otherwise.}

\item{alphaVar}{column name used for the transparency, empty by default.}

\item{alpha}{character or factor with specified transparency(s) for the bars,
replicated if needed. By default: '1' if alphaVar is not specified.}

\item{alphaRange}{transparency (alpha) range used in the plot, possible only
if the alphaVar is 'numeric' or 'integer'.}
}
\value{
ggplot object
}
\description{
Format aesthetics for the main plot with waterfall plot
}
\author{
Katarzyna Gorczak
}
