% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-topTables_extract.R
\name{formatTTcoef}
\alias{formatTTcoef}
\title{Format top table for a specific coefficient}
\usage{
formatTTcoef(
  topTableCoef,
  columns,
  mean,
  coefI,
  coef,
  featuresIdVar,
  hoverText
)
}
\arguments{
\item{topTableCoef}{top table for one coefficient}

\item{columns}{columns to extract}

\item{mean}{logical (FALSE by default) whether to extract column with 
averaged expression or logCPM ("AveExpr" or "logCPM") depending on the 
\code{input}}

\item{coefI}{coefficient of interest}

\item{coef}{character vector with coefficients}

\item{featuresIdVar}{column name with unique feature ids (empty by default)}

\item{hoverText}{logical whether to add point label for interactive plot
(FALSE by default)}
}
\value{
data.frame
}
\description{
Format top table for a specific coefficient
}
\author{
Laure Cougnaud, Katarzyna Gorczak
}
