% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-plot_functions.R
\name{formatVariableLength}
\alias{formatVariableLength}
\title{shorten string length}
\usage{
formatVariableLength(x, var, length)
}
\arguments{
\item{x}{data.frame with \code{var}}

\item{var}{column name}

\item{length}{max number of characters in each string}
}
\value{
string of length \code{length}

Character vector with shortened variable.
}
\description{
shorten string length
}
\author{
Katarzyna Gorczak
}
