% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-plot_functions.R
\name{formatVariableSpace}
\alias{formatVariableSpace}
\title{remove space in variable name}
\usage{
formatVariableSpace(var)
}
\arguments{
\item{var}{string}
}
\value{
String with formatted variable
}
\description{
remove space in variable name
}
