% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-plot_functions.R
\name{getCoefColor}
\alias{getCoefColor}
\title{Create colors for coefficient labels}
\usage{
getCoefColor(coef, coefLabel, coefColor)
}
\arguments{
\item{coef}{character, coefficient names.}

\item{coefLabel}{character vector of labels or a function to transform the 
existing labels.}

\item{coefColor}{color palette for coef labels.}
}
\value{
named character (named color palette)
}
\description{
Create colors for coefficient labels
}
\author{
Katarzyna Gorczak
}
