% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-checkCoefficients.R
\name{getCoefLabel}
\alias{getCoefLabel}
\title{Get coefficient label}
\usage{
getCoefLabel(coef, coefLabel)
}
\arguments{
\item{coef}{character vector with coefficients}

\item{coefLabel}{character vector of coefficient labels or a 
function to transform the coefficients}
}
\value{
labels of the coefficients; either a vector or a list 
with nested labels
}
\description{
Get coefficient label
}
