% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-plot_functions.R
\name{getFeatureColor}
\alias{getFeatureColor}
\title{Create colors for feature labels}
\usage{
getFeatureColor(
  topTableOutput,
  features,
  featuresIdVar,
  featuresColor,
  order = FALSE,
  plot = c("logRatio", "heatmap", "waterfall")
)
}
\arguments{
\item{topTableOutput}{combined topTables for all coefficients}

\item{features}{feature ids or numeric. If NULL (by default), 
show top 20 features according to the first \code{coef}. If numeric and 
\code{input} is a list, the tables are subsetted according to the row names
 of the table for the first \code{coef}.}

\item{featuresIdVar}{column name with feature ids.}

\item{featuresColor}{character vector specifying colors to use for 
the feature label text. Either length 1 or the same length as 
\code{features}, allowing particular features to be highlighted.}

\item{order}{logical whether to order features 
(used only when plot is 'logRatio')}

\item{plot}{plot}
}
\value{
named color palette for feature labels
}
\description{
Create colors for feature labels
}
\author{
Katarzyna Gorczak
}
