% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-getFeatures.R
\name{getFeatures}
\alias{getFeatures}
\title{Get features from input}
\usage{
getFeatures(input, featuresIdVar = character())
}
\arguments{
\item{input}{model or a list with top tables named with coefficients.\cr
For model: object of class \code{MArrayLM} 
(linear model, see \code{\link[limma]{eBayes}}), \code{DGELRT} 
(see \code{edgeR}) or \code{DESeqResults} (see \code{DESeq2}) are supported.
See more details on the top table format in section:
\link[=daVis-common-doc]{Top table format}.}

\item{featuresIdVar}{column with unique feature identifiers}
}
\value{
character with all features from input
}
\description{
Get features from input
}
\author{
Katarzyna Gorczak
}
