% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-checkCoefficients.R
\name{getInputIdCoef}
\alias{getInputIdCoef}
\title{Get ID of the input \code{list} matching to the specific coefficient}
\usage{
getInputIdCoef(input, coef)
}
\arguments{
\item{input}{input list}

\item{coef}{character vector of length 1 with coefficient}
}
\value{
numeric of length 1 specifying which element of the list
contains the coefficient
}
\description{
Get ID of the input \code{list} matching to the specific coefficient
}
