% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-checkCoefficients.R
\name{getModelCoefs}
\alias{getModelCoefs}
\title{Get coefficients in input}
\usage{
getModelCoefs(input, coef = NULL)
}
\arguments{
\item{input}{object of class \code{MArrayLM} (see \code{limma}), 
\code{DGELRT} (see \code{edgeR}) or \code{DESeqResults} (see \code{DESeq2})}

\item{coef}{(optional) coefficient of interest}
}
\value{
Character vector with coefficients
}
\description{
Get coefficients in input
}
