% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-plot_functions.R
\name{getNumberOfRegulatedGenes}
\alias{getNumberOfRegulatedGenes}
\title{Get number of significant genes for coefficient}
\usage{
getNumberOfRegulatedGenes(
  input,
  fdr = 0.05,
  logFCrange = NULL,
  dir = c("up", "down")
)
}
\arguments{
\item{input}{a list of top tables.}

\item{fdr}{threshold for adjusted p-value.}

\item{logFCrange}{numeric, upper and lower bound for logFC}

\item{dir}{direction, select either up- or down-regulated genes 
('up' - genes with logFC larger than 0, 
'down' - genes with logFC smaller than 0).}
}
\value{
named numeric vector with number of up- or down-regulated genes
}
\description{
Get number of significant genes for coefficient
}
\author{
Katarzyna Gorczak
}
