% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-plot_functions.R
\name{getNumberOfSignificantGenes}
\alias{getNumberOfSignificantGenes}
\title{Get number of significant genes for coefficient}
\usage{
getNumberOfSignificantGenes(input, logFCrange = NULL, fdr = 0.05)
}
\arguments{
\item{input}{a list of top tables.}

\item{logFCrange}{numeric, upper and lower bound for logFC}

\item{fdr}{threshold for adjusted p-value.}
}
\value{
named numeric vector with number of significant genes
}
\description{
Get number of significant genes for coefficient
}
\author{
Katarzyna Gorczak
}
