% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-getFeatures.R
\name{getTopFeatures}
\alias{getTopFeatures}
\title{Get top features}
\usage{
getTopFeatures(input, coef, featuresIdVar = character(), n = 10)
}
\arguments{
\item{input}{model or a list with top tables named with coefficients.\cr
For model: object of class \code{MArrayLM} 
(linear model, see \code{\link[limma]{eBayes}}), \code{DGELRT} 
(see \code{edgeR}) or \code{DESeqResults} (see \code{DESeq2}) are supported.
See more details on the top table format in section:
\link[=daVis-common-doc]{Top table format}.}

\item{coef}{character vector with coefficient}

\item{featuresIdVar}{column with unique feature identifiers}

\item{n}{numeric, number of top features}
}
\value{
top n features based on their adjusted p-value
}
\description{
Get top features
}
\author{
Laure Cougnaud, Katarzyna Gorczak
}
