% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-plot_functions.R
\name{getTopGenes}
\alias{getTopGenes}
\title{Get top genes with highest significance and top genes with highest logFC}
\usage{
getTopGenes(input, topGenes, featuresIdVar = character())
}
\arguments{
\item{input}{long-format table}

\item{topGenes}{number of top genes to extract}

\item{featuresIdVar}{column with feature identifier (must be unique)}
}
\value{
data.frame with top genes
}
\description{
Get top genes with highest significance and top genes with highest logFC
}
\author{
Laure Cougnaud, Katarzyna Gorczak
}
