% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-topTables_extract.R
\name{getTopTableFromModel}
\alias{getTopTableFromModel}
\title{Extract top tables from a model}
\usage{
getTopTableFromModel(input, coef, se = FALSE)
}
\arguments{
\item{input}{object of class \code{MArrayLM} (see \code{limma}), 
\code{DGELRT} (see \code{edgeR}) or \code{DESeqResults} (see \code{DESeq2})}

\item{coef}{coefficient of interest}

\item{se}{Logical (FALSE by default), should standard errors be extracted?}
}
\value{
data.frame with top tables for coefficient of interest.\cr
The features are ordered in original (unsorted) order in the model.\cr
Standard errors, if requested are in the column: 'SE'.
}
\description{
Extract top tables from a model
}
\author{
Katarzyna Gorczak
}
