% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-plotTheme.R
\name{ggPlotTheme}
\alias{ggPlotTheme}
\title{Customized ggplot theme}
\usage{
ggPlotTheme(
  ggObject,
  panelBackground = NULL,
  panelBorder = "black",
  title = NULL,
  titleSize = 2,
  xTextSize = 1.2,
  xTextColor = "black",
  xTitle = NULL,
  xTitleSize = 1.4,
  xTextAngle = 0,
  xTextHjust = 0.5,
  yTextSize = 1.2,
  yTextColor = "black",
  yTitle = NULL,
  yTitleSize = 1.4,
  facetLabelSize = 1.5,
  facetLabelColor = "black",
  gridMajor = TRUE,
  gridMinor = TRUE,
  yTicks = TRUE,
  xTicks = TRUE,
  legendPosition = "right",
  legendTitleSize = 1.3,
  legendTextSize = 1.1
)
}
\arguments{
\item{ggObject}{ggplot object}

\item{panelBackground}{panel background of the entire plot}

\item{panelBorder}{border around plotting area, by default black}

\item{title}{plot title}

\item{titleSize}{cex for plot title, relative to the parent}

\item{xTextSize}{cex for axis text}

\item{xTextColor}{color font on the x axis, by default black}

\item{xTitle}{title of the x axis}

\item{xTitleSize}{cex for the x axis title}

\item{xTextAngle}{angle of the x axis labels}

\item{xTextHjust}{numeric for horizontal adjustment, 0.5 by default}

\item{yTextSize}{cex for axis text}

\item{yTextColor}{color font on the y axis, by default black}

\item{yTitle}{title of the y axis, NULL by default}

\item{yTitleSize}{cex for the y axis title}

\item{facetLabelSize}{cex for the facet labels}

\item{facetLabelColor}{color of the text in facets}

\item{gridMajor}{logical whether to show the major grid lines}

\item{gridMinor}{logical whether to show the minor grid lines}

\item{yTicks}{logical whether to show ticks on the y axis}

\item{xTicks}{logical whether to show ticks on the x axis}

\item{legendPosition}{legend position, by default right}

\item{legendTitleSize}{cex for the legend title}

\item{legendTextSize}{cex for the legend}
}
\value{
ggplot object with customized theme
}
\description{
Customized ggplot theme
}
