% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-checkInput.R
\name{isModel}
\alias{isModel}
\title{Check if input is a model}
\usage{
isModel(input)
}
\arguments{
\item{input}{object of class \code{MArrayLM} (see \code{limma}), 
\code{DGELRT} (see \code{edgeR}) or \code{DESeqResults} (see \code{DESeq2})}
}
\value{
logical
}
\description{
Check if input is a model
}
