% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-checkInput.R
\name{isTopTable}
\alias{isTopTable}
\title{Check if input is a topTable containing colums with logFC, 
p-value and adjusted p-value}
\usage{
isTopTable(input)
}
\arguments{
\item{input}{a top table (see \code{limma::topTable})}
}
\value{
logical
}
\description{
Check if input is a topTable containing colums with logFC, 
p-value and adjusted p-value
}
