% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots-scatterPlot.R
\name{labelCorr}
\alias{labelCorr}
\title{Add label with correlation to scatter plot}
\usage{
labelCorr(topTableOutput, g, correlationCex)
}
\arguments{
\item{topTableOutput}{combined topTables for all coefficients}

\item{g}{ggplot object}

\item{correlationCex}{cex for the text with correlation value}
}
\value{
ggplot object
}
\description{
Add label with correlation to scatter plot
}
\author{
Katarzyna Gorczak
}
