% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-plot_functions.R
\name{labelGenesOfInterest}
\alias{labelGenesOfInterest}
\title{Add layer with genes of interest to ggplot object}
\usage{
labelGenesOfInterest(
  g,
  topTableOutputGenesOfInterest,
  typePlot,
  color = "red",
  sizeVar = character(),
  size = if (length(sizeVar) > 0) numeric() else 2,
  genesToHighlightCex,
  ...
)
}
\arguments{
\item{g}{ggplot object}

\item{topTableOutputGenesOfInterest}{data.frame with \code{genesToHighlight}}

\item{typePlot}{plot type can be one of "static" or "interactive".}

\item{color}{color palette, only used if \code{colorVar} is specified.}

\item{sizeVar}{column name used for the size, empty by default.}

\item{size}{character or factor with specified size(s) (cex) for the points, 
replicated if needed. This is used only if sizeVar is empty. 
By default: '2.5' if sizeVar is not specified and 
default ggplot size(s) otherwise}

\item{genesToHighlightCex}{cex for genesToHighlight}

\item{...}{Extra parameters passed to \code{geom_text_repel} to customize
the position of the gene labels.}
}
\value{
ggplot object
}
\description{
Add layer with genes of interest to ggplot object
}
\author{
Laure Cougnaud, Katarzyna Gorczak
}
