% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots-logRatioPlot.R
\name{labelTextLRP}
\alias{labelTextLRP}
\title{Add optional text to the log-ratio plot}
\usage{
labelTextLRP(topTableOutput, errorBars, g, textVar, textVarCex)
}
\arguments{
\item{topTableOutput}{combined topTables for all coefficients}

\item{errorBars}{logical whether to add error bars to the plot (+/- one
standard error). If \code{input} (or each element within a 
\code{input} list) is a:
\itemize{
\item{model: only supported for \code{MArrayLM}}
\item{top table: a 'se' column should be available}
}}

\item{g}{ggplot object with volcano plot}

\item{textVar}{(optional) String with name of a column to display as text.}

\item{textVarCex}{cex for the text next to bars}
}
\value{
ggplot object
}
\description{
Add optional text to the log-ratio plot
}
\author{
Laure Cougnaud, Kirsten Van Hoorde, Katarzyna Gorczak
}
