% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-plot_functions.R
\name{labelTopGenes}
\alias{labelTopGenes}
\title{Add point labels to ggplot object}
\usage{
labelTopGenes(
  g,
  includeTableGenesOfInterest,
  topTableOutputTopGenes,
  topTableOutputGenesOfInterest,
  colorVar,
  topGenesCex,
  ...
)
}
\arguments{
\item{g}{ggplot object}

\item{includeTableGenesOfInterest}{whether to label \code{genesToHighlight}}

\item{topTableOutputTopGenes}{data.frame with top genes}

\item{topTableOutputGenesOfInterest}{data.frame with \code{genesToHighlight}}

\item{colorVar}{string with column name containing variable used 
for coloring}

\item{topGenesCex}{cex for topGenes labels}

\item{...}{Extra parameters passed to \code{geom_text_repel} to customize
the position of the gene labels}
}
\value{
ggplot object
}
\description{
Add point labels to ggplot object
}
\author{
Laure Cougnaud, Katarzyna Gorczak
}
