% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots-heatmap.R
\name{mainH}
\alias{mainH}
\title{Create main plot object with heatmap}
\usage{
mainH(typePlot, topTableOutput, color, colorNA)
}
\arguments{
\item{typePlot}{plot type, can be one of "static" or "interactive".}

\item{topTableOutput}{combined topTables for all coefficients}

\item{color}{palette for gradient to fill the heatmap.}

\item{colorNA}{color for missing values.}
}
\value{
ggplot object
}
\description{
Create main plot object with heatmap
}
\author{
Katarzyna Gorczak
}
