% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots-logRatioPlot.R
\name{mainLRP}
\alias{mainLRP}
\title{Create main plot object with log-ratio plot}
\usage{
mainLRP(typePlot, color, topTableOutput, errorBars)
}
\arguments{
\item{typePlot}{plot type, can be one of "static" or "interactive".}

\item{color}{string with colors for the bars, one per coefficient.}

\item{topTableOutput}{combined topTables for all coefficients}

\item{errorBars}{logical whether to add error bars to the plot (+/- one
standard error). If \code{input} (or each element within a 
\code{input} list) is a:
\itemize{
\item{model: only supported for \code{MArrayLM}}
\item{top table: a 'se' column should be available}
}}
}
\value{
ggplot object
}
\description{
Create main plot object with log-ratio plot
}
\author{
Laure Cougnaud, Kirsten Van Hoorde, Katarzyna Gorczak
}
