% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots-MAplot.R
\name{mainMA}
\alias{mainMA}
\title{Create main plot object with MA plot}
\usage{
mainMA(typePlot, direction, color, sizeVar, topTableOutput, alpha, size)
}
\arguments{
\item{typePlot}{plot type can be one of "static" or "interactive".}

\item{direction}{logical whether to color significant up- and 
down-regulated genes}

\item{color}{colors for points indicating direction (should be three colors:
significant up- and down-regulated and non-significant)}

\item{sizeVar}{column name used for the size, empty by default.}

\item{topTableOutput}{combined topTables for all coefficients}

\item{alpha}{transparency level for the points, 0.4 by default.}

\item{size}{character or factor with specified size(s) (cex) for the points, 
replicated if needed. This is used only if sizeVar is empty. 
By default: '2.5' if sizeVar is not specified and 
default ggplot size(s) otherwise}
}
\value{
ggplot object
}
\description{
Create main plot object with MA plot
}
\author{
Katarzyna Gorczak
}
