% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots-scatterPlot.R
\name{mainSP}
\alias{mainSP}
\title{Create main plot object with scatter plot}
\usage{
mainSP(
  typePlot,
  color,
  topTableOutput,
  xlab,
  coef,
  ylab,
  multiplePlot,
  pointSize,
  alpha
)
}
\arguments{
\item{typePlot}{plot type, can be one of "static" or "interactive"}

\item{color}{color palette to distinguish significance groups. 
Four colors must be specified.}

\item{topTableOutput}{combined topTables for all coefficients}

\item{xlab}{x-axis title, NULL by default.}

\item{coef}{character, coefficient names.}

\item{ylab}{y-axis title, NULL by default.}

\item{multiplePlot}{logical whether to use facet on coefficients}

\item{pointSize}{point size, 2 by default.}

\item{alpha}{transparency level for the points, 0.4 by default.}
}
\value{
ggplot object
}
\description{
Create main plot object with scatter plot
}
\author{
Katarzyna Gorczak
}
