% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots-volcanoPlot.R
\name{mainVP}
\alias{mainVP}
\title{Create main plot object with volcano plot}
\usage{
mainVP(
  colorVar,
  shapeVar,
  alphaVar,
  sizeVar,
  topTableOutput,
  typePlot,
  color,
  shape,
  alpha,
  size
)
}
\arguments{
\item{colorVar}{(optional) string with column name containing variable used 
for coloring. If not specified, coloring is based on the adjusted p-values.}

\item{shapeVar}{(optional) string with column name containing variable used 
for shaping. If not specified, only points are used.}

\item{alphaVar}{column name used for the transparency, empty by default.}

\item{sizeVar}{column name used for the size, empty by default.}

\item{topTableOutput}{combined topTables for all coefficients}

\item{typePlot}{plot type can be one of "static" or "interactive".}

\item{color}{color palette, only used if \code{colorVar} is specified.}

\item{shape}{shape palette, only used if \code{shapeVar} is specified.}

\item{alpha}{character or factor with specified transparency(s) for the 
points, replicated if needed. By default: '1' if alphaVar is not specified.}

\item{size}{character or factor with specified size(s) (cex) for the points, 
replicated if needed. This is used only if sizeVar is empty. 
By default: '2.5' if sizeVar is not specified and 
default ggplot size(s) otherwise}
}
\value{
ggplot object
}
\description{
Create main plot object with volcano plot
}
\author{
Laure Cougnaud, Kirsten Van Hoorde, Katarzyna Gorczak
}
