% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots-waterfallPlot.R
\name{mainWP}
\alias{mainWP}
\title{Create main plot object with waterfall plot}
\usage{
mainWP(
  typePlot,
  colorVar,
  fillVar,
  alphaVar,
  topTableOutput,
  fill,
  color,
  alpha
)
}
\arguments{
\item{typePlot}{plot type, can be one of "static" or "interactive".}

\item{colorVar}{name of variable (in 'input') used for coloring, 
empty by default}

\item{fillVar}{name of variable (in 'input') used for filling, 
empty by default}

\item{alphaVar}{column name used for the transparency, empty by default.}

\item{topTableOutput}{combined topTables for all coefficients}

\item{fill}{character or factor with specified color(s) for the boxplot 
inside, replicated if needed. By default: 'skyblue2' if fillVar is not 
specified and default ggplot palette otherwise.}

\item{color}{character or factor with specified color(s) for the bar border, 
replicated if needed. By default: 'skyblue2' if colorVar is not specified 
and default ggplot palette otherwise.}

\item{alpha}{character or factor with specified transparency(s) for the bars,
replicated if needed. By default: '1' if alphaVar is not specified.}
}
\value{
ggplot object
}
\description{
Create main plot object with waterfall plot
}
\author{
Katarzyna Gorczak
}
