% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-topTables_arrange.R
\name{makeElementsUnique}
\alias{makeElementsUnique}
\title{Make elements unique}
\usage{
makeElementsUnique(x)
}
\arguments{
\item{x}{vector with values}
}
\value{
vector with unique values, 
for duplicated values a suffix '_' is added
}
\description{
Make elements unique
}
