% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-exampleData.R
\name{modelExampleData}
\alias{modelExampleData}
\title{Normalize data, create model matrix and contrasts}
\usage{
modelExampleData(eset)
}
\arguments{
\item{eset}{ExpressionSet}
}
\value{
a list with data, model, design matrix and contrast matrix
}
\description{
Normalize data, create model matrix and contrasts
}
\author{
Katarzyna Gorczak
}
