% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-topTables_extract.R
\name{orderComparison}
\alias{orderComparison}
\title{Sort the comparison columns according to the coefficients}
\usage{
orderComparison(df, coefLabel)
}
\arguments{
\item{df}{top table for one coefficient}

\item{coefLabel}{character vector with coefficient labels or list with
such vector (in case of multiple labels)}
}
\value{
data.frame
}
\description{
Sort the comparison columns according to the coefficients
}
\author{
Laure Cougnaud, Katarzyna Gorczak
}
