% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-plot_functions.R
\name{processFeatures}
\alias{processFeatures}
\title{Process table for visualizations.}
\usage{
processFeatures(tbl, coef, featuresIdVar, order)
}
\arguments{
\item{tbl}{logn-form table (top tables for specified coefs)}

\item{coef}{character, coefficient names.}

\item{featuresIdVar}{column name with feature ids.}

\item{order}{character vector with feature ids in a specific order}
}
\value{
data.frame with processed features
}
\description{
Some can be missing across different coefficients (either different subsets
of to tables are provided or provided 'features' are not present in each top
table). The table will be processed for each coefficient: 
\itemize{
\item{1: missing features will be added (if a missing feature in coef X is
present in coef Y, this label will be used)}
\item{2: feature missing in all coefs will obviously not be plotted}
\item{3: the same 'featuresVar' can happen for different 'featuresIdVar' - 
add index to duplicates as 'featuresVar' is used for plotting}
}
}
\author{
Katarzyna Gorczak
}
