% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-topTables_arrange.R
\name{reshapeTable}
\alias{reshapeTable}
\title{Reshape table}
\usage{
reshapeTable(table, timevar)
}
\arguments{
\item{table}{long- or wide-format table}

\item{timevar}{column name in long-format table that differentiates multiple
records from the same group}
}
\value{
a list of tables
}
\description{
Reshape table
}
\author{
Katarzyna Gorczak
}
