% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-exampleData.R
\name{runDESeq2}
\alias{runDESeq2}
\title{Fit the model with DESeq2}
\usage{
runDESeq2(input)
}
\arguments{
\item{input}{list of objects (data, contrasts and model matrix)}
}
\value{
DESeq2 output from results() function
}
\description{
Fit the model with DESeq2
}
\author{
Katarzyna Gorczak
}
