% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-exampleData.R
\name{runEdgeR}
\alias{runEdgeR}
\title{Fit the model with edgeR}
\usage{
runEdgeR(input)
}
\arguments{
\item{input}{list of objects (data, contrasts and model matrix)}
}
\value{
edgeR output from glmQLFTest() function
}
\description{
Fit the model with edgeR
}
\author{
Katarzyna Gorczak
}
