% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-exampleData.R
\name{runLimma}
\alias{runLimma}
\title{Fit the model with limma}
\usage{
runLimma(input)
}
\arguments{
\item{input}{list of objects (data, contrasts and model matrix)}
}
\value{
limma outpur from eBayes() function
}
\description{
Fit the model with limma
}
\author{
Katarzyna Gorczak
}
