% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-exampleData.R
\name{runTopTable}
\alias{runTopTable}
\title{Fit the model and return top table}
\usage{
runTopTable(input)
}
\arguments{
\item{input}{list of objects (data, contrasts and model matrix)}
}
\value{
a list of top tables
}
\description{
Fit the model and return top table
}
\author{
Katarzyna Gorczak
}
