% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-plot_functions.R
\name{setCategoricalElement}
\alias{setCategoricalElement}
\title{check if variable is not numeric (or integer)}
\usage{
setCategoricalElement(x, typeVar)
}
\arguments{
\item{x}{data.frame}

\item{typeVar}{column name in \code{x}}
}
\value{
Logical, TRUE if the variable is not a numeric or integer
}
\description{
check if variable is not numeric (or integer)
}
