% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-plot_functions.R
\name{setGradientScale}
\alias{setGradientScale}
\title{check if manual aesthetic for the gradient should be set}
\usage{
setGradientScale(x, typeVar, valVar)
}
\arguments{
\item{x}{data.frame with \code{typeVar}}

\item{typeVar}{name of variable for aesthetic}

\item{valVar}{fixed value of variable of aesthetic}
}
\value{
logical, if TRUE the manual scale should be set
}
\description{
This is the case only if \code{typeVar} and \code{valVar} are specified,
and if the variable is numeric or integer
}
\author{
Katarzyna Gorczak
}
