% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-getFeatures.R
\name{subsetFeatures}
\alias{subsetFeatures}
\title{Subset input with specific features}
\usage{
subsetFeatures(input, features, featuresIdVar = character())
}
\arguments{
\item{input}{model or a list with top tables named with coefficients.\cr
For model: object of class \code{MArrayLM} 
(linear model, see \code{\link[limma]{eBayes}}), \code{DGELRT} 
(see \code{edgeR}) or \code{DESeqResults} (see \code{DESeq2}) are supported.
See more details on the top table format in section:
\link[=daVis-common-doc]{Top table format}.}

\item{features}{features of interest}

\item{featuresIdVar}{column with unique feature identifiers}
}
\value{
input with subset of features
}
\description{
Subset input with specific features
}
\author{
Katarzyna Gorczak
}
