% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-getFeatures.R
\name{topFeaturesDeseq}
\alias{topFeaturesDeseq}
\title{Extract top features from \code{DESeq2} output}
\usage{
topFeaturesDeseq(input, n, featuresIdVar)
}
\arguments{
\item{input}{model or a list with top tables named with coefficients.\cr
For model: object of class \code{MArrayLM} 
(linear model, see \code{\link[limma]{eBayes}}), \code{DGELRT} 
(see \code{edgeR}) or \code{DESeqResults} (see \code{DESeq2}) are supported.
See more details on the top table format in section:
\link[=daVis-common-doc]{Top table format}.}

\item{n}{numeric, number of top features}

\item{featuresIdVar}{column with unique feature identifiers}
}
\value{
character vector with feature names
}
\description{
Extract top features from \code{DESeq2} output
}
\author{
Katarzyna Gorczak
}
