% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots-logRatioPlot.R
\name{daLogRatioPlot}
\alias{daLogRatioPlot}
\title{Plot log ratios}
\usage{
daLogRatioPlot(
  input,
  coef = NULL,
  coefLabel = NULL,
  features = NULL,
  featuresIdVar = character(),
  featuresVar = featuresIdVar,
  featuresOrder = NULL,
  featuresMaxNChar = 50,
  featuresColor = "black",
  color = character(),
  text = NULL,
  textCex = 4,
  xlab,
  xexpand = if (!is.null(text)) {
     ggplot2::expansion(mult = 0.2)
 },
  axesCex = 1,
  axesTitleCex = 1,
  title = NULL,
  titleCex = 1,
  facetCex = 1,
  facetColor = "black",
  facetNCol = NULL,
  errorBars = TRUE,
  typePlot = c("static", "interactive")
)
}
\arguments{
\item{input}{model or a list with top tables named with coefficients.\cr
For model: object of class \code{MArrayLM} 
(linear model, see \code{\link[limma]{eBayes}}), \code{DGELRT} 
(see \code{edgeR}) or \code{DESeqResults} (see \code{DESeq2}) are supported.
See more details on the top table format in section:
\link[=daVis-common-doc]{Top table format}.}

\item{coef}{character, coefficient names.}

\item{coefLabel}{character vector of labels or function to transform 
the existing labels, or list with such labels.
\cr If a list is specified, nested facets are used. In that case, 
the coefficients are ordered based on the groups defined by the coefficient 
labels.}

\item{features}{feature ids or numeric. If NULL (by default), 
show top 20 features according to the first \code{coef}. If numeric and 
\code{input} is a list, the tables are subsetted according to the row names
 of the table for the first \code{coef}.}

\item{featuresIdVar}{column name with feature ids.}

\item{featuresVar}{column name with feature ids to label.}

\item{featuresOrder}{if not NULL, features in the graph are re-ordered, 
either based on similarity ('similarity') or significance ('significance') 
of the statistics. See section 'Feature ordering'.}

\item{featuresMaxNChar}{maximum number of characters to truncate the 
feature labels to.}

\item{featuresColor}{character vector specifying colors to use for 
the feature label text. Either length 1 or the same length as 
\code{features}, allowing particular features to be highlighted.}

\item{color}{string with colors for the bars, one per coefficient.}

\item{text}{(optional) String with name of a column, or function to extract
a text from the columns of a top table, to display as text.\cr
See the available columns in the top table in section:
\link[=daVis-common-doc]{Top table format}.}

\item{textCex}{cex for the text next to bars}

\item{xlab}{x-axis title, or \code{NULL} to remove.\cr
By default, set to: 'logFC' or 'logFC (+- SE)' if error bar(s)
are available.}

\item{xexpand}{expansion factor for the x-axis 
(see \code{\link[ggplot2]{expansion}}. If \code{text} is specified, 
the x-axis is expanded by 20\% on each side.}

\item{axesCex}{relative size for the label of the axes.}

\item{axesTitleCex}{relative size for the title of the axes.}

\item{title}{plot title, NULL by default.}

\item{titleCex}{relative size for the title of the plot.}

\item{facetCex}{cex for the facets if multiple plot are used.}

\item{facetColor}{color for the text of the facets}

\item{facetNCol}{number of columns in facets, by default the function 
\code{n2mfrow} is used to.}

\item{errorBars}{logical whether to add error bars to the plot (+/- one
standard error). If \code{input} (or each element within a 
\code{input} list) is a:
\itemize{
\item{model: only supported for \code{MArrayLM}}
\item{top table: a 'se' column should be available}
}}

\item{typePlot}{plot type, can be one of "static" or "interactive".}
}
\value{
ggplot object
}
\description{
This is a function to create barplot with log-foldchanges.
}
\section{Feature ordering}{

The features are ordered based on:
\itemize{
\item{'similarity': a hierarchical clustering
of the (Euclidean) distances between the statistics. The statistics are for:
\itemize{
\item{limma: the t-statistic of each coefficient}
\item{edgeR: the (overall) F-statistic is considered}
}}
\item{'significance': decreasing average (-log10) p-values
across coefficients. The p-values are for:
\itemize{
\item{limma: the significance of each coefficient}
\item{edgeR: the (overall) p-value of the model}
}}
}
}

\examples{
exampleData <- createExampleData(path = ".", output = c("limma", "topTable"))
model <- exampleData$limma
topTableList <- exampleData$topTable
coefs <- c("B.LvsP", "L.LvsP", "B.PvsV", "L.PvsV")

# Simple log-ratio plot
daLogRatioPlot(input = model, coef = coefs, facetNCol = 4)

# Specify features and annotation
features <- daVis:::getTopFeatures(input = model, coef = "B.LvsP", 
featuresIdVar = "ENTREZID", n = 20)
daLogRatioPlot(input = model, coef = coefs, 
features = features, featuresIdVar = "ENTREZID", 
featuresVar = c("SYMBOL", "GENENAME"), featuresMaxNChar = 35, facetNCol = 4)

# Specify different set of features
coefs <- c("B.LvsP", "L.LvsP", "B.PvsV", "L.PvsV", "A")
set.seed(123)
features <- sample(features, 20)
daLogRatioPlot(
input = list(model, A = topTableList[["B.LvsP"]][c(seq_len(6), 9, 10), ]),
featuresIdVar = "ENTREZID", features = features, coef = coefs, 
facetNCol = 5, errorBars = TRUE)

# Sort features as specified
coefs <- c("B.LvsP", "L.LvsP", "B.PvsV", "L.PvsV")
daLogRatioPlot(input = model, featuresIdVar = "ENTREZID", 
features = features, coef = coefs, facetNCol = 4, errorBars = TRUE)

# Sort features based on similarity
daLogRatioPlot(input = model, featuresIdVar = "ENTREZID", 
features = features, coef = coefs, facetNCol = 4, errorBars = TRUE,
featuresOrder = "similarity")

# see vignette for other examples

}
\author{
Laure Cougnaud, Katarzyna Gorczak, Heather Turner, 
Aditya Bhagwat, Kirsten Van Hoorde
}
