% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-topTables_extract.R
\name{getSEModel}
\alias{getSEModel}
\title{Get standard error from a limma model.}
\usage{
getSEModel(input, coef)
}
\arguments{
\item{input}{object of class \code{MArrayLM} (see \code{limma})}

\item{coef}{character vector with one single coefficient}
}
\value{
If \code{input} is \code{MArrayLM}, a numeric vector with SE named
by feature ID, NA otherwise.
}
\description{
The standard errors are computed, according to BioC support 
\href{https://support.bioconductor.org/p/70175/#70185}{question 70175}, 
as: \deqn{stdev.unscaled * \sqrt{s2.post}}
with: 
\itemize{
\item{stdev.unscaled: unscaled standard deviations}
\item{s2.post: posterior values for sigma^2}
}
}
\examples{
exampleData <- createExampleData(path = ".", output = "limma")
model <- exampleData$limma
getSEModel(input = model, coef = "B.LvsP")
}
\author{
Laure Cougnaud
}
