% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots-logRatioPlot.R
\name{orderFeatures}
\alias{orderFeatures}
\title{Order features based on similarity or significance}
\usage{
orderFeatures(input, featuresOrder, featuresIdVar)
}
\arguments{
\item{input}{list of top tables}

\item{featuresOrder}{if not NULL, features in the graph are re-ordered, 
either based on similarity ('similarity') or significance ('significance') 
of the statistics. See section 'Feature ordering'.}

\item{featuresIdVar}{column name with feature ids.}
}
\value{
ordered features
}
\description{
Order features based on similarity or significance
}
\section{Feature ordering}{

The features are ordered based on:
\itemize{
\item{'similarity': }{a hierarchical clustering
of the (Euclidean) distances between the statistics. The statistics are for:
\itemize{
\item{limma: }{the t-statistic of each coefficient}
\item{edgeR: }{the (overall) F-statistic is considered}
}}
\item{'significance': }{decreasing average (-log10) p-values
across coefficients. The p-values are for:
\itemize{
\item{limma: }{the significance of each coefficient}
\item{edgeR: }{the (overall) p-value of the model}
}}
}
}

